% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyte_pathway_repository.R
\docType{class}
\name{AnalytePathwayRepository-class}
\alias{AnalytePathwayRepository-class}
\alias{AnalytePathwayRepository}
\title{A Reference class for querying the analytehaspathway table}
\arguments{
\item{analyte_ids}{list of analyte ids used to filter results}

\item{pathway_ramp_ids}{list of pathway ids used to filter results
Method for retrieving pathwayRampId from analytehaspathway table}

\item{distinct}{TRUE if you only want unique values}

\item{in_pathway_sources}{list of sources used to filter results}

\item{not_in_pathway_sources}{list of sources ids used to filter results}
}
\description{
A Reference class for querying the analytehaspathway table
}
\section{Fields}{

\describe{
\item{\code{db_connection}}{a DBConnection object
Method for retrieving all columns from analytehaspathway table}
}}


