% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyte_repository.R
\docType{class}
\name{AnalyteRepository-class}
\alias{AnalyteRepository-class}
\alias{AnalyteRepository}
\title{A Reference class for querying analytes in the database}
\arguments{
\item{analyte_ids}{list of analyte ids used to filter results}

\item{analyte_names}{list of analyte names used to filter results}
}
\description{
A Reference class for querying analytes in the database
}
\section{Fields}{

\describe{
\item{\code{db_connection}}{a DBConnection object
Method for retrieving analytes}
}}


