% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{bhCorrect}
\alias{bhCorrect}
\title{Helper function to return a Benjamini-Hochberg (BH) corrected p-values. The method supports
chemical class enrichment p-value corrections.}
\usage{
bhCorrect(resultMat)
}
\arguments{
\item{resultMat}{is a dataframe containing p-values in a column named 'p-value'}
}
\value{
the input dataFrame with BH corrected p-values with column name 'adjP_BH'
}
\description{
Helper function to return a Benjamini-Hochberg (BH) corrected p-values. The method supports
chemical class enrichment p-value corrections.
}
