% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{buildAnalyteOverlapPerRxnLevelUpsetDataframe}
\alias{buildAnalyteOverlapPerRxnLevelUpsetDataframe}
\title{Creates the input dataframe for the upset plot created in 'plotAnalyteOverlapPerRxnLevel'}
\usage{
buildAnalyteOverlapPerRxnLevelUpsetDataframe(
  reactionsResults,
  includeCofactorMets = FALSE
)
}
\arguments{
\item{reactionClassesResults}{output of getReactionClassesForAnalytes()}
}
\description{
Creates the input dataframe for the upset plot created in 'plotAnalyteOverlapPerRxnLevel'
}
