% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{buildChemicalClassDataframe}
\alias{buildChemicalClassDataframe}
\title{Creates the input dataframe for the interactive plot created in 'plotChemicalClass'}
\usage{
buildChemicalClassDataframe(chemicalClassResults)
}
\arguments{
\item{chemicalClassResults}{output of getChemClass()}
}
\description{
Creates the input dataframe for the interactive plot created in 'plotChemicalClass'
}
