% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{buildChemicalClassSurveryDataframe}
\alias{buildChemicalClassSurveryDataframe}
\title{Creates the input dataframe for the interactive plot created in 'plotChemicalClassSurvery'}
\usage{
buildChemicalClassSurveryDataframe(chemicalClassSurveryResults)
}
\arguments{
\item{chemicalClassSurveryResults}{output of getReactionClassesForAnalytes()}
}
\description{
Creates the input dataframe for the interactive plot created in 'plotChemicalClassSurvery'
}
