% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{buildFrequencyTables}
\alias{buildFrequencyTables}
\title{Return all analytes that map to unique pathways in a pathwaydf}
\usage{
buildFrequencyTables(inputdf)
}
\arguments{
\item{inputdf}{internal df with pathwayramp ids}
}
\value{
dataframe of all analytes that map to the input pathways
}
\description{
Return all analytes that map to unique pathways in a pathwaydf
}
\author{
Andrew Christopher Patt
}
