% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampChemPropQueries.R
\name{buildPropertyList}
\alias{buildPropertyList}
\title{Internal function to validate property list}
\usage{
buildPropertyList(propList, db = RaMP())
}
\arguments{
\item{propList}{an optional list of specific properties to extract.  Options include 'all' (default),  'iso_smiles', 'inchi_key', 'inchi_key_prefix', 'inchi', 'mw', 'monoisotop_mass', 'formula', 'common_name'.}

\item{db}{a RaMP database object}
}
\description{
Internal function to validate property list
}
