% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{buildReactionClassesSunburstDatafarme}
\alias{buildReactionClassesSunburstDatafarme}
\title{Creates the input dataframe for the sunburst plot created in 'plotReactionClasses'}
\usage{
buildReactionClassesSunburstDatafarme(reactionClassesResults = "")
}
\arguments{
\item{reactionClassesResults}{output of getReactionClassesForAnalytes()}
}
\description{
Creates the input dataframe for the sunburst plot created in 'plotReactionClasses'
}
