% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{buildReactionClassesSunburstDataframe}
\alias{buildReactionClassesSunburstDataframe}
\title{Creates the input dataframe for the sunburst plot created in 'plotReactionClasses'}
\usage{
buildReactionClassesSunburstDataframe(reactionClassesResults)
}
\arguments{
\item{reactionClassesResults}{output of getReactionClassesForAnalytes()}
}
\description{
Creates the input dataframe for the sunburst plot created in 'plotReactionClasses'
}
