% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{checkIdPrefixes}
\alias{checkIdPrefixes}
\title{Check for id prefixes
By default, the function checks on whether there are at least 90% of input user analytes with
appropriate RaMP-supported prefixes.}
\usage{
checkIdPrefixes(idList, perc_cutoff = 0.9)
}
\arguments{
\item{idList}{list of ids (typically input by user)}

\item{perc_cutoff}{percent cutof used to throw a warning that many ids are missing}
}
\description{
Check for id prefixes
By default, the function checks on whether there are at least 90% of input user analytes with
appropriate RaMP-supported prefixes.
}
