% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampChemClassQueries.R
\name{chemicalClassEnrichment}
\alias{chemicalClassEnrichment}
\title{Returns chemical class information comparing a metabolite subset to a metabolite population, including Fisher Exact Test
enrichment p-values and FDR values.}
\usage{
chemicalClassEnrichment(mets, pop = "database")
}
\arguments{
\item{mets}{a vector of source prepended metabolite ids}

\item{pop}{an optional vector of source prepended metabolite ids to be used as the background reference population of
metabolites for enrichment. If "database", the background population is taken as all RaMP DB metabolites. Default: 'database'}
}
\value{
a list of dataframes, each holding chemical classs enrichment statistics for specific chemical classification systems,
such as HMDB Classyfire class categories and LIPIDMAPS class categories.  The results list chemical classes, metabolite hits counts,
Fisher Exact p-values and Benjamini-Hochberg corrected p-values (FDR estimates)
}
\description{
Returns chemical class information comparing a metabolite subset to a metabolite population, including Fisher Exact Test
enrichment p-values and FDR values.
}
\examples{
\dontrun{
# metabolite list of interest
metList = c('hmdb:HMDB0000056',
            'hmdb:HMDB0000439',
            'hmdb:HMDB0000479',
            'hmdb:HMDB0000532',
            'hmdb:HMDB0001015',
            'hmdb:HMDB0001138',
            'hmdb:HMDB0029159',
            'hmdb:HMDB0029412',
            'hmdb:HMDB0034365',
            'hmdb:HMDB0035227',
            'hmdb:HMDB0007973',
            'hmdb:HMDB0008057',
            'hmdb:HMDB0011211')

# the background population can be a separate ID list (preferred) or all database entries
# (skip pop parameter).
pkg.globals <- setConnectionToRaMP(dbname="ramp2",username="root",conpass="",host = "localhost")
enrichedClassStats <- chemicalClassEnrichment(mets = metList)
}
}
