% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{chemicalClassSurveyRampIdsConn}
\alias{chemicalClassSurveyRampIdsConn}
\title{chemicalClassSurveyRampIdsConn is a helper function that takes a list of metabolite ids, a list of 'population' metabolite ids
and a MariaDB Connection object. The method returns metabolite class information for the metabolite list and a population of all ramp metabolites.}
\usage{
chemicalClassSurveyRampIdsConn(mets, pop, conn)
}
\arguments{
\item{mets}{a list object of prefixed metabolite ids of interest}

\item{pop}{a list object of prefixed metabolite ids, representing a larger population of metabolites from which the mets were selected.}

\item{conn}{a MariaDB Connection object to support queries}
}
\value{
a list object containing three objects 'count_summary', 'met_classes' and 'met_query_report'.
The count_summary is a dataframe containing metabolite classes and number of metabolites in each class.
The met_classes is a detailed listing of compound classes associated with each input metabolite
The met_query_report indicates the number of input metabolites, how many were found in the DB and the list of metabolites not found in RaMP DB.
}
\description{
chemicalClassSurveyRampIdsConn is a helper function that takes a list of metabolite ids, a list of 'population' metabolite ids
and a MariaDB Connection object. The method returns metabolite class information for the metabolite list and a population of all ramp metabolites.
}
