% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printingFunctions.R
\name{cleanup}
\alias{cleanup}
\title{Generic function for removing RaMP identifiers and making query results more pleasant to look at}
\usage{
cleanup(data, show_n_rows = 6)
}
\arguments{
\item{data}{A dataframe from the following functions: runCominedFishersTests, getAnalyteFromPathway, getPathwayFromAnalyte, chemicalClassSurvey, and getChemicalProperties}

\item{show_n_rows}{Only print first n rows of input (Shows all if NA)}
}
\value{
Printed out and prettified version of input
}
\description{
Generic function for removing RaMP identifiers and making query results more pleasant to look at
}
\author{
Andrew Christopher Patt
}
