% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{common_to_ramp}
\alias{common_to_ramp}
\title{Find rampId from given common name vector
The rampId can be plugged in other functions to continue query}
\usage{
common_to_ramp(
  common_names,
  conpass = NULL,
  dbname = "ramp",
  username = "root",
  host = "localhost",
  socket = NULL
)
}
\arguments{
\item{common_names}{a vector of metabolite common name strings}

\item{conpass}{password for database access (string)}

\item{dbname}{name of the mysql database (default is "ramp")}

\item{username}{username for database access (default is "root")}

\item{host}{host name for database access (default is "localhost")}

\item{socket}{(optional) location of mysql.sock file}
}
\value{
data.frame that has rampId as column
}
\description{
Find rampId from given common name vector
The rampId can be plugged in other functions to continue query
}
