% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapMatrix.R
\name{compute_overlap_matrix}
\alias{compute_overlap_matrix}
\title{Compute overlaping matrix based on given list return by findAnalyteHasPathway()}
\usage{
compute_overlap_matrix(pathwayid, pathwaysWithAnalytes, methods)
}
\arguments{
\item{pathwayid}{a vector that has all ramp pathway id in the pathwaysWithAnalytes}

\item{pathwaysWithAnalytes}{a list that has pathways ramp Id as name, analytes (ramp compound id only
or gene id only or both) as content. Return by findAnalyteHasPathway}

\item{methods}{must be in c('balanced','weighted') to determine which way to calculate this matrix}
}
\value{
the overlap matrix that has the overlap
}
\description{
Compute overlaping matrix based on given list return by findAnalyteHasPathway()
}
