% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectToRaMP.R
\name{connectToRaMP}
\alias{connectToRaMP}
\title{Connect to RaMP database (requires mysql password when running locally)}
\usage{
connectToRaMP()
}
\value{
MySQL connection based on given dbname,username,password, and host
}
\description{
Connect to RaMP database (requires mysql password when running locally)
}
\examples{
\dontrun{
pkg.globals <- setConnectionToRaMP(dbname="ramp2",username="root",conpass="",host = "localhost")
con <- connectToRaMP()
}
}
