% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printingFunctions.R
\name{display_results}
\alias{display_results}
\title{Generic function for removing RaMP identifiers and making query results more pleasant to look at}
\usage{
display_results(data, truncate_cells_at = 20, show_n_rows = 6)
}
\arguments{
\item{data}{A dataframe from the following functions: runCominedFishersTests, getAnalyteFromPathway, getPathwayFromAnalyte, chemicalClassSurvey, and getChemicalProperties}

\item{truncate_cells_at}{String length to truncate cells at (No truncation if NA)}

\item{show_n_rows}{Only print first n rows of input (Shows all if NA)}
}
\value{
Printed out and prettified version of input
}
\description{
Generic function for removing RaMP identifiers and making query results more pleasant to look at
}
\author{
Andrew Christopher Patt
}
