% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab3.R
\name{FilterFishersResults}
\alias{FilterFishersResults}
\title{Filter pathways by p-value cutoff for display and clustering}
\usage{
FilterFishersResults(fishers_df, p_cutoff)
}
\arguments{
\item{fishers_df}{The data frame generated by runFisherTest}

\item{p_cutoff}{p value cutoff determined by the user}
}
\description{
Filter pathways by p-value cutoff for display and clustering
}
