% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{filterPathwaysByAnalyteCount}
\alias{filterPathwaysByAnalyteCount}
\title{filterPathwaysByAnalyteCount utility method filtered a dataframe based on the number of analytes associated with rampPathwayIds contained in the dataframe.
Like fisher exact code, this one retains pathways with analyte count >= min_path_size, and having analyte_count < max_path_size}
\usage{
filterPathwaysByAnalyteCount(
  db = RaMP(),
  pathway_dataframe,
  pathway_ramp_id_col_name = "pathwayRampId",
  min_path_size = 5,
  max_path_size = 150
)
}
\arguments{
\item{db}{a RaMP databse object}

\item{pathway_dataframe}{a dataframe containing at least one column that contains rampPathwayIds}

\item{pathway_ramp_id_col_name}{the column name containing the rampPathwayIds}

\item{min_path_size}{the minimum number of pathway members (genes and metabolites) to include the pathway in the output (default = 5)}

\item{max_path_size}{the maximum number of pathway memnbers (genes and metaboltes) to include the pathway in the output (default = 150)}
}
\description{
filterPathwaysByAnalyteCount utility method filtered a dataframe based on the number of analytes associated with rampPathwayIds contained in the dataframe.
Like fisher exact code, this one retains pathways with analyte count >= min_path_size, and having analyte_count < max_path_size
}
