% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnPathwaysEnrich_InputAnalytes.R
\name{findCluster}
\alias{findCluster}
\title{Perform fuzzy multiple linkage partitioning clustering on pathways
identified by Fisher's test}
\usage{
findCluster(
  fishersDf,
  percAnalyteOverlap = 0.5,
  minPathwayToCluster = 2,
  percPathwayOverlap = 0.5,
  db = RaMP(),
  ...
)
}
\arguments{
\item{fishersDf}{The full result object generated by runEnrichPathways}

\item{percAnalyteOverlap}{Minimum overlap for pathways to be considered similar
(Default = 0.5)}

\item{minPathwayToCluster}{Minimum number of 'similar' pathways required to start
a cluster (medoid) (Default = 3)}

\item{percPathwayOverlap}{Minimum overlap for clusters to merge (Default = 0.5)}

\item{db}{a RaMP database object}

\item{...}{Internal Use - for handling deprecated parameter names}
}
\value{
list:[[1]] Pathway enrichment result with dataframe having a cluster assignment column added
[[2]] analyte type
[[3]] cluster assignment in the list form
}
\description{
Perform fuzzy multiple linkage partitioning clustering on pathways
identified by Fisher's test
}
\examples{
\dontrun{
pathways.enriched <- runEnrichPathways(
analytes = c("hmdb:HMDB0000033","hmdb:HMDB0000052","hmdb:HMDB0000094",
             "hmdb:HMDB0000161","hmdb:HMDB0000168","hmdb:HMDB0000191","hmdb:HMDB0000201",
             "chemspider:10026","hmdb:HMDB0006059", "Chemspider:6405", "CAS:5657-19-2",
             "hmdb:HMDB0002511", "chemspider:20171375","CAS:133-32-4", "CAS:5746-90-7",
             "CAS:477251-67-5", "hmdb:HMDB0000695", "chebi:15934", "CAS:838-07-3",
             "hmdb:HMDBP00789", "hmdb:HMDBP00283", "hmdb:HMDBP00284", "hmdb:HMDBP00850"),
db=rampDB)

filtered.pathways.enriched <- filterEnrichResults(enrichResults=pathways.enriched,
                                                  pValType = 'holm', pValCutoff=0.05)

clusters <- findCluster(filtered.pathways.enriched, percAnalyteOverlap = 0.2,
                        percPathwayOverlap = 0.2, db=rampDB)
}
}
