% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{findDuplicatePathways}
\alias{findDuplicatePathways}
\title{Return list of duplicate Wikipathway IDs from Reactome. This may be unnecessary in the future}
\usage{
findDuplicatePathways(db = RaMP())
}
\arguments{
\item{db}{a RaMP database object}
}
\value{
List of duplicate Wikipathway IDs from Reactome.
}
\description{
Return list of duplicate Wikipathway IDs from Reactome. This may be unnecessary in the future
}
\author{
John Braisted
}
