% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_results_clustering_service.R
\docType{class}
\name{FisherResultsClustering-class}
\alias{FisherResultsClustering-class}
\alias{FisherResultsClustering}
\title{Reference class for clustering results from fisher's test}
\arguments{
\item{perc_analyte_overlap}{Minimum overlap for pathways to be considered similar
(Default = 0.5)}

\item{min_pathway_tocluster}{Minimum number of 'similar' pathways required to start
a cluster (medoid) (Default = 3)}

\item{perc_pathway_overlap}{Minimum overlap for clusters to merge (Default = 0.5)}
}
\value{
list:[[1]] Pathway enrichment result dataframe with cluster assignment column added
[[2]] analyte type
[[3]] cluster assignment in the list form
}
\description{
Reference class for clustering results from fisher's test
}
\section{Fields}{

\describe{
\item{\code{fishers_results}}{fisher's test results}

\item{\code{fishers_results_df}}{dataframe set during test method}

\item{\code{cluster_list}}{cluster_list}

\item{\code{analyte_type}}{analyte type}

\item{\code{pathway_matrix}}{dataframe with pathways matrix}
}}


\examples{
\dontrun{
pathwaydf<-getPathwayFromAnalyte(c("MDM2","TP53","glutamate","creatinine"),
                NameOrIds="names", conpass=conpass)
fisher.results <- runCombinedFisherTest(pathwaydf=pathwaydf,conpass=conpass)
filtered.fisher.results <- FilterFishersResults(fisher.results,p_holmadj_cutoff=0.05)
filteredclust.fisher.results <- findCluster(filtered.fisher.results)
}
}
