% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_test_service.R
\docType{class}
\name{FisherTest-class}
\alias{FisherTest-class}
\alias{FisherTest}
\title{Reference class for performing fisher's test}
\arguments{
\item{pathwaydf}{a data frame resulting from getPathwayFromAnalyte}

\item{total_metabolites}{number of metabolites analyzed in the experiment (e.g. background) (default is 1000; set to 'NULL' to retrieve total number of metabolites that map to any pathway in RaMP). Assumption that analyte_type is "metabolite")}

\item{total_genes}{number of genes analyzed in the experiment (e.g. background) (default is 20000, with assumption that analyte_type is "genes")}

\item{analyte_type}{"metabolites" or "genes" (default is "metabolites")}

\item{conpass}{password for database access (string)}

\item{dbname}{name of the mysql database (default is "ramp")}

\item{username}{username for database access (default is "root")}

\item{host}{host name for database access (default is "localhost")}

\item{perc_analyte_overlap}{Minimum overlap for pathways to be considered similar
(Default = 0.5)}

\item{min_pathway_tocluster}{Minimum number of 'similar' pathways required to start
a cluster (medoid) (Default = 3)}

\item{perc_pathway_overlap}{Minimum overlap for clusters to merge (Default = 0.5)}

\item{fishers_df}{The data frame generated by runFisherTest}

\item{p_holmadj_cutoff}{return pathways where Holm adjusted pvalues are < p_holmadj_cutoff}

\item{p_fdradj_cutoff}{return pathways where FDR adjusted pvalues are < p_fdradj_cutoff}
}
\value{
a dataframe with columns containing pathway ID, fisher's p value, user analytes in pathway, and total analytes in pathway
Perform fuzzy multiple linkage partitioning clustering on pathways identified by
Fisher's test

list:[[1]] Pathway enrichment result dataframe with cluster assignment column added
[[2]] analyte type
[[3]] cluster assignment in the list form

list:[[1]]Dataframe with pathway enrichment results, only significant pathways
[[2]]analyte type
}
\description{
Reference class for performing fisher's test
}
\section{Fields}{

\describe{
\item{\code{db_connection}}{DBConnection object}

\item{\code{fishresults}}{dataframe set during test method}

\item{\code{analyte_type}}{analyte type}
}}


\examples{
\dontrun{
pathwaydf<-getPathwayFromAnalyte(c("MDM2","TP53","glutamate","creatinine"),
                NameOrIds="names", conpass=conpass)
fisher.results <- runCombinedFisherTest(pathwaydf=pathwaydf,conpass=conpass)
filtered.fisher.results <- FilterFishersResults(fisher.results,p_holmadj_cutoff=0.05)
filteredclust.fisher.results <- findCluster(filtered.fisher.results)
}
\dontrun{
pathwaydf<-getPathwayFromAnalyte(c("MDM2","TP53","glutamate","creatinine"),
                NameOrIds="names", conpass=conpass)
fisher.results <- runCombinedFisherTest(pathwaydf=pathwaydf,conpass=conpass)
filtered.fisher.results <- FilterFishersResults(fisher.results,p_holmadj_cutoff=0.05)
}
}
