% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaMP.R
\name{get_remote_db_version_list}
\alias{get_remote_db_version_list}
\title{List RaMP db versions}
\usage{
get_remote_db_version_list(branch = "main")
}
\arguments{
\item{branch}{Which GitHub branch to pull a list of remote databases from}
}
\description{
returns the list of RaMP db versions available
}
