% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceDataFunctions.R
\name{getCurrentRaMPDBVersions}
\alias{getCurrentRaMPDBVersions}
\title{Retrieve versions of each database within the current version of RaMP}
\usage{
getCurrentRaMPDBVersions()
}
\value{
database source version info
}
\description{
Retrieve versions of each database within the current version of RaMP
}
\examples{
\dontrun{
pkg.globals <- setConnectionToRaMP(dbname="ramp2",username="root",conpass="",host = "localhost")
getCurrentRaMPDBVersions()
}
}
