% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceDataFunctions.R
\name{getCurrentRaMPVersion}
\alias{getCurrentRaMPVersion}
\title{Retrieve RaMP version}
\usage{
getCurrentRaMPVersion()
}
\value{
current ramp databse version
}
\description{
Retrieve RaMP version
}
\examples{
\dontrun{
pkg.globals <- setConnectionToRaMP(dbname="ramp2",username="root",conpass="",host = "localhost")
getCurrentRaMPVersion()
}
}
