% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnPathwaysEnrich_InputAnalytes.R
\name{getCustomPathwayFromAnalyte}
\alias{getCustomPathwayFromAnalyte}
\title{Utility method supporting pathway analyses when file-based pathway lists are used rather than the ramp database.}
\usage{
getCustomPathwayFromAnalyte(analytes, pathways, analyte_type)
}
\arguments{
\item{analytes}{a vector of analytes (genes or metabolites) that need to be searched}

\item{pathways}{If "RaMP" (default), use pathway definitions within RaMP-DB. Else, supply path to gmx file containing custom pathway definitions. GMX files are a tab-separated format that contain one analyte set per column, with the name of the set in the first row, and constituent analytes in subsequent rows}

\item{analyte_type}{"genes" or "metabolites"}
}
\value{
A pathwaydf compatible with runFisherTest
}
\description{
Utility method supporting pathway analyses when file-based pathway lists are used rather than the ramp database.
}
\author{
Andrew Patt
}
