% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceDataFunctions.R
\name{getEntityCountsFromSourceDBs}
\alias{getEntityCountsFromSourceDBs}
\title{Retrieve counts of entitites (e.g. Metabolites, Pathways, Metabolite-Pathway associations, etc.) for RaMP source databases}
\usage{
getEntityCountsFromSourceDBs()
}
\value{
database sources and entity counts associated with each data source
}
\description{
Retrieve counts of entitites (e.g. Metabolites, Pathways, Metabolite-Pathway associations, etc.) for RaMP source databases
}
\examples{
\dontrun{
pkg.globals <- setConnectionToRaMP(dbname="ramp2",username="root",conpass="",host = "localhost")
getEntityCountsFromSourceDBs()
}
}
