% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metClassUtil.R
\name{getMetabClassDataSources}
\alias{getMetabClassDataSources}
\title{Returns class data sources for metabolites}
\usage{
getMetabClassDataSources(db = RaMP())
}
\arguments{
\item{db}{a RaMP database object, if not specified a new one is created with RaMP::RaMP()}
}
\value{
Returns list of data sources for metabolites
}
\description{
Returns class data sources for metabolites
}
