% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metClassUtil.R
\name{getOntologies}
\alias{getOntologies}
\title{Returns list of available ontologies from database}
\usage{
getOntologies(db = RaMP())
}
\arguments{
\item{db}{a RaMP database object, if not specified a new one is created with RaMP::RaMP()}
}
\value{
Returns ontologies listed in database
}
\description{
Returns list of available ontologies from database
}
