% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceDataFunctions.R
\name{getPathwayNameList}
\alias{getPathwayNameList}
\title{Retrieve list of pathway names}
\usage{
getPathwayNameList(db = RaMP())
}
\arguments{
\item{db}{a RaMP database object}
}
\value{
vector of unique pathway names (alphabetically ordered)
}
\description{
Retrieve list of pathway names
}
\examples{
\dontrun{
pkg.globals <- setConnectionToRaMP(dbname="ramp2",username="root",conpass="",host = "localhost")
getPathwayNameList()
}
}
