% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampReactionQueries.R
\name{getProteinsForReactions}
\alias{getProteinsForReactions}
\title{getProteinsForReactions returns protein information for a list of reaction ids.
This utility methed can help extend information from previous queries.
For instance, if a user queries for reactions related to a list of metabolites,
this method can be used to return proteins on some subset of reaction ids to find related proteins.}
\usage{
getProteinsForReactions(db = RaMP(), reactionList = c())
}
\arguments{
\item{db}{a RaMP databse object}

\item{reactionList}{list of reaction ids}
}
\value{
returns a dataframe object with reaction to protein mappings.
}
\description{
getProteinsForReactions returns protein information for a list of reaction ids.
This utility methed can help extend information from previous queries.
For instance, if a user queries for reactions related to a list of metabolites,
this method can be used to return proteins on some subset of reaction ids to find related proteins.
}
