% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceDataFunctions.R
\name{getRaMPAnalyteIntersections}
\alias{getRaMPAnalyteIntersections}
\title{Retrieve RaMP Analyte Source Intersections, these indicate the level of analyte overlaps between our sources}
\usage{
getRaMPAnalyteIntersections(analyteType = "metabolites", format = "json")
}
\arguments{
\item{analyteType}{returns analyte overlaps for 'metabolites' or 'genes'}

\item{format}{can be one of either 'json', 'upsetR_expression'}
}
\value{
current analyte overlaps counts between current data sources, for the specified analyteType
}
\description{
Retrieve RaMP Analyte Source Intersections, these indicate the level of analyte overlaps between our sources
}
\examples{
\dontrun{
pkg.globals <- setConnectionToRaMP(dbname="ramp2",username="root",conpass="",host = "localhost")
jsonResult <- getRaMPAnalyteIntersections(analyteType='genes', format='json')
}
}
