% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlWorker.R
\name{getRampDbType}
\alias{getRampDbType}
\title{A user utility method that returns the current DB Type}
\usage{
getRampDbType()
}
\value{
MySQL or SQLite
}
\description{
A user utility method that returns the current DB Type
}
