% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampReactionQueries.R
\name{getRampSourceInfoFromAnalyteIDs}
\alias{getRampSourceInfoFromAnalyteIDs}
\title{getRampSourceInfoFromAnalyteIDs Utility method to extract source table information from analyte ids}
\usage{
getRampSourceInfoFromAnalyteIDs(db = RaMP(), analytes)
}
\arguments{
\item{analytes}{list of analyte ids}
}
\value{
returns a dataframe of ramp analyte source information
}
\description{
getRampSourceInfoFromAnalyteIDs Utility method to extract source table information from analyte ids
}
