% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampReactionQueries.R
\name{getReactionClassesForAnalytes}
\alias{getReactionClassesForAnalytes}
\title{getReactionClassesForAnalytes returns reactions class and EC numbers for a collection of input compound ids}
\usage{
getReactionClassesForAnalytes(
  db = RaMP(),
  analytes,
  multiRxnParticipantCount = 1,
  humanProtein = TRUE,
  concatResults = F,
  includeReactionIDs = F
)
}
\arguments{
\item{db}{a RaMP database object}

\item{analytes}{list of analyte ids}

\item{multiRxnParticipantCount}{minimum number of analytes to report a reaction class, default = 1}

\item{humanProtein}{require reactions to have a human protein (enzyme or transporter), default True}
}
\value{
returns a three dataframes of reaction EC classe information, one for each EC level
}
\description{
getReactionClassesForAnalytes returns reactions class and EC numbers for a collection of input compound ids
}
