% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampReactionQueries.R
\name{getReactionClassStats}
\alias{getReactionClassStats}
\title{Utility method that returns the number of rhea reaction count in each reaction class.}
\usage{
getReactionClassStats(db = RaMP())
}
\arguments{
\item{reactions}{input result from the getReactionsForAnalytes}
}
\description{
Utility method that returns the number of rhea reaction count in each reaction class.
}
