% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampReactionQueries.R
\name{getReactionDetails}
\alias{getReactionDetails}
\title{getReactionDetails returns general reaction information for a list of reaction ids.
This utility methed can help extend information from previous queries.
For instance, if a user queries for reactions related to a list of analytes, or filtered on reactions,
this method can be used to return general reaction info on some subset of reaction ids of interest.}
\usage{
getReactionDetails(reactionList = c(), db = RaMP())
}
\arguments{
\item{reactionList}{list of reaction ids}

\item{db}{a RaMP database object}
}
\value{
returns a dataframe object with reaction information.
}
\description{
getReactionDetails returns general reaction information for a list of reaction ids.
This utility methed can help extend information from previous queries.
For instance, if a user queries for reactions related to a list of analytes, or filtered on reactions,
this method can be used to return general reaction info on some subset of reaction ids of interest.
}
