% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampReactionQueries.R
\name{getReactionParticipants}
\alias{getReactionParticipants}
\title{getReactionParticipants returns protein information for a list of reaction ids.
This utility method can help extend information from previous queries.
For instance, if a user queries for reactions related to a list of metabolites,
this method can be used to return proteins on some subset of reaction ids to find related proteins.}
\usage{
getReactionParticipants(reactionList = c(), db = RaMP())
}
\arguments{
\item{reactionList}{list of reaction ids}

\item{db}{a RaMP databse object}
}
\value{
returns a dataframe object with reaction to reaction participant mappings.
}
\description{
getReactionParticipants returns protein information for a list of reaction ids.
This utility method can help extend information from previous queries.
For instance, if a user queries for reactions related to a list of metabolites,
this method can be used to return proteins on some subset of reaction ids to find related proteins.
}
