% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampReactionQueries.R
\name{getReactionParticpantCounts}
\alias{getReactionParticpantCounts}
\title{Utility method that evalutates the mapping counts for analytes to reaction ids}
\usage{
getReactionParticpantCounts(analyte2Rxn, minRxnParticipantCountFilter = 1)
}
\arguments{
\item{analyte2Rxn}{result object from getReactionsForAnalytes}

\item{minRxnParticipantCountFilter}{if > 1, the set of reactions is reduced to those with having this number of mapped analytes}
}
\description{
Utility method that evalutates the mapping counts for analytes to reaction ids
}
