% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampReactionQueries.R
\name{getReactionRheaURLs}
\alias{getReactionRheaURLs}
\title{Utility method that returns rhea reaction urls}
\usage{
getReactionRheaURLs(reactionList)
}
\arguments{
\item{reactionList}{list of prefixed (rhea:) reaction ids}
}
\description{
Utility method that returns rhea reaction urls
}
