% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampReactionQueries.R
\name{getReactionsForAnalytes}
\alias{getReactionsForAnalytes}
\title{getReactionsForAnalytes}
\usage{
getReactionsForAnalytes(
  analytes,
  analyteType = "metabolites",
  namesOrIds = "ids",
  onlyHuman = T
)
}
\arguments{
\item{analytes}{list of analytes}

\item{analyteType}{analyte type, 'metabolites' (default), 'genes' or 'both'}

\item{namesOrIds}{indicates if input analyte list contains identifiers or analyte names}

\item{onlyHuman}{boolean to only return pathways with only human metabolites and human enzymes (dev in progress...)}
}
\value{
a dataframe of reaction information on each input analyte
}
\description{
getReactionsForAnalytes
}
