% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampReactionQueries.R
\name{getReactionsForRaMPCompoundIds}
\alias{getReactionsForRaMPCompoundIds}
\title{getReactionsForRaMPCompoundIds returns reactions for a collection of RaMP compound ids}
\usage{
getReactionsForRaMPCompoundIds(
  db = RaMP(),
  rampCompoundIds,
  onlyHumanMets = F,
  humanProtein = F,
  includeTransportRxns = F,
  rxnDirs = c("UN")
)
}
\arguments{
\item{db}{RaMP object}

\item{rampCompoundIds}{list of ramp compound ids}

\item{onlyHumanMets}{boolean to only return pathways containing only human metabolites (ChEBI ontology) (dev in progress)}

\item{humanProtein}{boolean to only control pathways catalyzed by a human proteins (having human Uniprot) (dev in progress)}

\item{includeTransportRxns}{if TRUE, returns metabolic and transport reactions}
}
\value{
returns a dataframe of reaction information for each ramp compound id
}
\description{
getReactionsForRaMPCompoundIds returns reactions for a collection of RaMP compound ids
}
