% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampReactionQueries.R
\name{getReactionSourceIdsFromReactionQuery}
\alias{getReactionSourceIdsFromReactionQuery}
\title{Utility method that returns the reaction source ids for a given result from getReactionsForAnalytes}
\usage{
getReactionSourceIdsFromReactionQuery(reactions)
}
\arguments{
\item{reactions}{input result from the getReactionsForAnalytes}
}
\description{
Utility method that returns the reaction source ids for a given result from getReactionsForAnalytes
}
