% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampReactionQueries.R
\name{getRheaAnalyteReactionAssociations}
\alias{getRheaAnalyteReactionAssociations}
\title{getAnalyteReactionAssociations returns mets associated with input proteins, and proteins associated with input metabolites
Associations are made through shared Rhea reactions}
\usage{
getRheaAnalyteReactionAssociations(
  analytes,
  includeRheaRxnDetails = F,
  humanProtein = TRUE,
  db = RaMP()
)
}
\arguments{
\item{analytes}{list of analyte ids}

\item{includeRheaRxnDetails}{returns additional columns that includes info in the reaction connecting mets and proteins}

\item{humanProtein}{require reactions to have a human protein (enzyme or transporter), default True}

\item{db}{a RaMP database object}
}
\value{
returns a three dataframes of reaction EC classe information, one for each EC level
}
\description{
getAnalyteReactionAssociations returns mets associated with input proteins, and proteins associated with input metabolites
Associations are made through shared Rhea reactions
}
