% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampReactionQueries.R
\name{getRheaMetabolitesForProteins}
\alias{getRheaMetabolitesForProteins}
\title{getRheaEnzymesAndTransportersForMetabolites returns proteins that are either enzymes or transporters
that are reaction participants with the input metabolite ids.}
\usage{
getRheaMetabolitesForProteins(
  analytes,
  includeRheaRxnDetails = F,
  humanProtein = T,
  db = RaMP()
)
}
\arguments{
\item{analytes}{analyte id vector}

\item{includeRheaRxnDetails}{returns additional columns that inlucdes info in the reaction connecting mets and proteins}

\item{humanProtein}{require reactions to have a human protein (enzyme or transporter), default True}

\item{db}{a RaMP database object}
}
\value{
returns a dataframe object with input metabolites, and reaction and associated protein informatin.
}
\description{
getRheaEnzymesAndTransportersForMetabolites returns proteins that are either enzymes or transporters
that are reaction participants with the input metabolite ids.
}
