% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{getTotalFoundInCategories}
\alias{getTotalFoundInCategories}
\title{Utility method to return metabolite counts found in compound class categories
based on an input data compound class data object from the chemicalClassSurvey function
The returned counts for each class category are for both the metabolite id query list
and for the larger full or user-defined population of metabolite ids.
This method is used in the exported chemicalClassEnrichment function}
\usage{
getTotalFoundInCategories(classData, inferIdMapping = FALSE)
}
\arguments{
\item{classData}{Data object returned from a call to chemicalClassSurvey
This input contains lists of chemical classes that pertain to a query list of metabolites and pertaining to
metabolites in a larger metabolite population.}
}
\value{
a list object with two keys, 'mets' and 'pop' that each has a table of metabolite or population
chemical classes and metabolite counts per class. This supports the chemicalClassEnrichment function.
}
\description{
Utility method to return metabolite counts found in compound class categories
based on an input data compound class data object from the chemicalClassSurvey function
The returned counts for each class category are for both the metabolite id query list
and for the larger full or user-defined population of metabolite ids.
This method is used in the exported chemicalClassEnrichment function
}
