% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralFunctions.R
\name{killDbConnections}
\alias{killDbConnections}
\title{Kill all databases connection.}
\usage{
killDbConnections()
}
\description{
If the connections is forgotten to be disconnected, it will cause potential trouble
for further connection.
}
