% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaMP.R
\name{listAvailableRaMPDbVersions}
\alias{listAvailableRaMPDbVersions}
\title{Lists local and remotely available RaMP SQLite DB versions and prompts with
message to download a new version if one exists.}
\usage{
listAvailableRaMPDbVersions(branch = "main")
}
\arguments{
\item{branch}{`character(1)` specifying the RaMP repository branch to load
the database from. by default (`version = "main"`) for the released versions
of the database.}
}
\description{
Lists local and remotely available RaMP SQLite DB versions and prompts with
message to download a new version if one exists.
}
