% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{listToQueryString}
\alias{listToQueryString}
\title{listToQueryString utility method to convert an id list to a comma separate string, with single quoted values.}
\usage{
listToQueryString(ids)
}
\arguments{
\item{ids}{list of ids (can be names or ids)}
}
\value{
comma separated list of single quoted ids or names
}
\description{
listToQueryString utility method to convert an id list to a comma separate string, with single quoted values.
}
