% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralFunctions.R
\name{loadOverlapMatrices}
\alias{loadOverlapMatrices}
\title{Load pathway overlap matrices for find_clusters function}
\usage{
loadOverlapMatrices()
}
\value{
A list of pathway overlap matrices for clustering
}
\description{
Load pathway overlap matrices for find_clusters function
}
