% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{ontologyEnrichmentResultsPlot}
\alias{ontologyEnrichmentResultsPlot}
\title{Cluster and plot significant ontologies by FDR-adjusted pval}
\usage{
ontologyEnrichmentResultsPlot(
  ontologiesSig,
  pval = "FDR",
  text_size = 8,
  sig_cutoff = 0.05,
  interactive = FALSE,
  db = RaMP()
)
}
\arguments{
\item{ontologiesSig}{output of FilterFishersResults}

\item{pval}{Which p value to plot, choose from Raw, FDR or Holm-adjusted}

\item{text_size}{Scales all text in figure (Default=16)}

\item{sig_cutoff}{Aesthetic, shows pvalue cutoff for significant ontologies}

\item{interactive}{If TRUE, return interactive plotly object instead of ggplot object}

\item{db}{a RaMP database object}
}
\description{
Cluster and plot significant ontologies by FDR-adjusted pval
}
