% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment_analysis.R
\name{pathway_enrichment_analysis}
\alias{pathway_enrichment_analysis}
\title{Pathway enrichment analysis}
\usage{
pathway_enrichment_analysis(
  analyte = "",
  identifier_type = "names",
  p_holmadj_cutoff = 0.05,
  p_fdradj_cutoff = NULL,
  perc_analyte_overlap = 0.2,
  perc_pathway_overlap = 0.2,
  min_pathway_tocluster = 2
)
}
\arguments{
\item{analyte}{vector of analyte names or Id's}

\item{identifier_type}{names or ids}

\item{p_holmadj_cutoff}{p_holmadj_cutoff}

\item{p_fdradj_cutoff}{p_fdradj_cutoff}

\item{perc_analyte_overlap}{perc_analyte_overlap}

\item{perc_pathway_overlap}{perc_pathway_overlap}

\item{min_pathway_tocluster}{min_pathway_tocluster}
}
\value{
list with pathway enrichment analysis results
}
\description{
Pathway enrichment analysis
}
