% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{pathway_ramp_to_source}
\alias{pathway_ramp_to_source}
\title{Convert pathway ramp ids to source ids}
\usage{
pathway_ramp_to_source(
  pathways,
  conpass = NULL,
  dbname = "ramp",
  username = "root",
  host = "localhost",
  socket = NULL
)
}
\arguments{
\item{pathways}{a vector of pathway source ids}

\item{conpass}{password for database access (string)}

\item{dbname}{name of the mysql database (default is "ramp")}

\item{username}{username for database access (default is "root")}

\item{host}{host name for database access (default is "localhost")}

\item{socket}{(optional) location of mysql.sock file}
}
\value{
vector of pathway source ids
}
\description{
Convert pathway ramp ids to source ids
}
