% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnPathwaysEnrich_InputAnalytes.R
\name{pathway_results_plot}
\alias{pathway_results_plot}
\title{Cluster and plot significant pathways by FDR-adjusted pval}
\usage{
pathway_results_plot(
  pathwaysSig,
  pval = "FDR",
  analyte_overlap_thresh = 0.2,
  pathway_overlap_thresh = 0.2,
  text_size = 16,
  sig_cutoff = 0.05,
  interactive = FALSE
)
}
\arguments{
\item{pathwaysSig}{output of FilterFisherResults}

\item{pval}{Which p value to plot, choose from Raw, FDR or Holm-adjusted}

\item{analyte_overlap_thresh}{Minimum overlap for pathways to be considered similar
(Default = 0.2)}

\item{pathway_overlap_thresh}{Minimum overlap for clusters to merge (Default = 0.2)}

\item{text_size}{Scales all text in figure (Default=16)}

\item{sig_cutoff}{Aesthetic, shows pvalue cutoff for significant pathways}

\item{interactive}{If TRUE, return interactive plotly object instead of ggplot object}
}
\description{
Cluster and plot significant pathways by FDR-adjusted pval
}
