% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{pathwayResultsPlot}
\alias{pathwayResultsPlot}
\title{Cluster and plot significant pathways by FDR-adjusted pval}
\usage{
pathwayResultsPlot(
  pathwaysSig,
  pval = "FDR",
  perc_analyte_overlap = 0.5,
  perc_pathway_overlap = 0.5,
  min_pathway_tocluster = 3,
  text_size = 8,
  sig_cutoff = 0.05,
  interactive = FALSE,
  db = RaMP()
)
}
\arguments{
\item{pathwaysSig}{output of FilterFisherResults}

\item{pval}{Which p value to plot, choose from Raw, FDR or Holm-adjusted}

\item{perc_analyte_overlap}{Minimum overlap for pathways to be considered similar
(Default = 0.2)}

\item{perc_pathway_overlap}{Minimum overlap for clusters to merge (Default = 0.2)}

\item{min_pathway_tocluster}{Minimum number of 'similar' pathways required to start
a cluster (medoid) (Default = 3)}

\item{text_size}{Scales all text in figure (Default=16)}

\item{sig_cutoff}{Aesthetic, shows pvalue cutoff for significant pathways}

\item{interactive}{If TRUE, return interactive plotly object instead of ggplot object}

\item{db}{a RaMP database object}
}
\description{
Cluster and plot significant pathways by FDR-adjusted pval
}
\examples{
\dontrun{
plotEnrichedPathways(filtered.fisher.results, text_size = 8, perc_analyte_overlap = 0.2,
                    min_pathway_tocluster = 2, perc_pathway_overlap = 0.2, interactive = FALSE, db = rampDB )
}
}
