% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotAnalyteOverlapPerRxnLevel}
\alias{plotAnalyteOverlapPerRxnLevel}
\title{Plots an interactive upset plot of overlapping input compounds at reaction class level 1}
\usage{
plotAnalyteOverlapPerRxnLevel(
  reactionsResults = "",
  includeCofactorMets = FALSE
)
}
\arguments{
\item{reactionsResults}{output of getReactionsForAnalytes()}

\item{includeCofactorMets}{include metabolites labeled at cofactors within ChEBI (Default = FALSE)}
}
\value{
An interactive HTML upset plot that allows the user to visualize the overlap in the number of input compounds across level 1 of reaction classes.
}
\description{
Plots an interactive upset plot of overlapping input compounds at reaction class level 1
}
