% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotChemicalClass}
\alias{plotChemicalClass}
\title{Plots an interactive sunburst plot of chemical classes}
\usage{
plotChemicalClass(chemicalClassResults, plotType = "sunburst")
}
\arguments{
\item{chemicalClassResults}{output of getChemClass()}

\item{plotType}{choice of 'sunburst' or 'treemap' plot type (default = 'sunburst')}
}
\value{
An interactive HTML sunburst plot that allows the user to pan/zoom into reaction classes of interest.
}
\description{
Plots an interactive sunburst plot of chemical classes
}
