% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotChemicalClassSurvery}
\alias{plotChemicalClassSurvery}
\title{Plots an interactive sunburst plot of chemical classes}
\usage{
plotChemicalClassSurvery(chemicalClassSurveryResults, plotType = "sunburst")
}
\arguments{
\item{chemicalClassSurveryResults}{output of chemicalClassSurvey()}

\item{plotType}{choice of 'sunburst' or 'treemap' plot type (default = 'sunburst')}
}
\value{
An interactive HTML sunburst plot that allows the user to pan/zoom into reaction classes of interest.
}
\description{
Plots an interactive sunburst plot of chemical classes
}
