% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotPathwayResults}
\alias{plotPathwayResults}
\title{Cluster and plot significant pathways by FDR-adjusted pVal}
\usage{
plotPathwayResults(
  pathwaysSig,
  pVal = "FDR",
  percAnalyteOverlap = 0.5,
  percPathwayOverlap = 0.5,
  minPathwayToCluster = 3,
  textSize = 8,
  sigCutoff = 0.05,
  interactive = FALSE,
  db = RaMP()
)
}
\arguments{
\item{pathwaysSig}{output of FilterFishersResults}

\item{pVal}{Which p value to plot, choose from Raw, FDR or Holm-adjusted}

\item{percAnalyteOverlap}{Minimum overlap for pathways to be considered similar
(Default = 0.2)}

\item{percPathwayOverlap}{Minimum overlap for clusters to merge (Default = 0.2)}

\item{minPathwayToCluster}{Minimum number of 'similar' pathways required to start
a cluster (medoid) (Default = 3)}

\item{textSize}{Scales all text in figure (Default=16)}

\item{sigCutoff}{Aesthetic, shows pvalue cutoff for significant pathways}

\item{interactive}{If TRUE, return interactive plotly object instead of ggplot object}

\item{db}{a RaMP database object}
}
\description{
Cluster and plot significant pathways by FDR-adjusted pVal
}
\examples{
\dontrun{
plotPathwayResults(pathwaysSig = filtered.fisher.results, textSize = 8, percAnalyteOverlap = 0.2,
   minPathwayToCluster = 2, percPathwayOverlap = 0.2, interactive = FALSE, db = rampDB )
}
}
