% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotReactionClasses}
\alias{plotReactionClasses}
\title{Plots an interactive sunburst plot of reaction class}
\usage{
plotReactionClasses(reactionClassesResults = "")
}
\arguments{
\item{reactionClassesResults}{output of getReactionClassesForAnalytes()}
}
\value{
An interactive HTML sunburst plot that allows the user to pan/zoom into reaction classes of interest.
}
\description{
Plots an interactive sunburst plot of reaction class
}
