% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processNewRamp.R
\name{processData}
\alias{processData}
\title{processData function generates pathway RampId frequency (gene or metabolite) based on pathway source (hmdb,kegg,reactome,wiki)}
\usage{
processData(conpass, dbname = "ramp", username = "root", host = "localhost")
}
\arguments{
\item{conpass}{a string that specifies password for database connection}

\item{dbname}{a string that specifies database name of MySQL database}

\item{username}{a string that specifies name of MySQL database}

\item{host}{a string that specifes host for database connection}
}
\value{
R object (FT_data.Rdata) with dataframes (hmdb_metab,hmdb_gene,kegg_gene,kegg_metab,reactome_gene,reactome_metab,wiki_gene,wiki_metab)
}
\description{
processData function generates pathway RampId frequency (gene or metabolite) based on pathway source (hmdb,kegg,reactome,wiki)
}
