% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{queryReport}
\alias{queryReport}
\title{queryReport is a helper function to report on the number of query items that were found and missed, and the list of missed query values.}
\usage{
queryReport(queryList, foundList)
}
\arguments{
\item{queryList}{is a list object that contains all user input query values}

\item{foundList}{is a list object of all user input query values that were retrieved during the query}
}
\value{
returns a list object with three return values, 'query_list_size', 'found_list_size', 'missed_query_elements'
The 'size' values are integers for the size of the input query and the number of input query values found.
The missed_query_elements is a list containing the subset of query values that are not found during the query.
}
\description{
queryReport is a helper function to report on the number of query items that were found and missed, and the list of missed query values.
}
