% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralFunctions.R
\name{rampCataOut}
\alias{rampCataOut}
\title{The function send query to databases to find synonym}
\usage{
rampCataOut(synonym, maxItems = 1000, con)
}
\arguments{
\item{synonym}{string value that describe synonym of metabolites}

\item{maxItems}{Integer value that describe maximum returned by
query}

\item{con}{a connection object returned from the function connectToRaMP()}
}
\value{
a dataframe that has all searching result with given synonym
as column name. If \code{synonym} is compound, it returns all gene that
catalyze that compound. If \code{synonym} is gene, it returns all compound
catalyzed by this gene.
}
\description{
based on catalyzation relationship. 
This function also automate identification about if the given
metabolites is gene or compound
}
