% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab5.R
\name{rampFastBiofluid}
\alias{rampFastBiofluid}
\title{Search by fast algorithm to find analyte or biofluid location}
\usage{
rampFastBiofluid(string, analyteOrBiofluid = NULL)
}
\arguments{
\item{string}{a string or data.frame that contains all given analyte or
biofluid location}

\item{analyteOrBiofluid}{a string or NULL value defined output format}
}
\value{
if analyteOrBiofluid == NULL, it returns a list, otherwise return
a data.frame that has analyte or biofluid
}
\description{
Search by fast algorithm to find analyte or biofluid location
}
