% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab4.R
\name{rampFastCata}
\alias{rampFastCata}
\title{Retrieves analytes that involved in same reaction as input metabolite}
\usage{
rampFastCata(analytes = NULL, conpass = NULL, dbname = "ramp",
  username = "root", host = "localhost", NameOrIds = "ids")
}
\arguments{
\item{analytes}{a vector of analytes (metabolites) that need to be searched}

\item{conpass}{password for database access (string)}

\item{dbname}{name of the mysql database (default is "ramp")}

\item{username}{username for database access (default is "root")}

\item{host}{host name for database access (default is "localhost")}

\item{NameOrIds}{whether input is "names" or "ids" (default is "ids")}
}
\value{
a dataframe that contains search results
}
\description{
Retrieves analytes that involved in same reaction as input metabolite
}
\examples{
\dontrun{
rampFastCata(analytes="creatine",conpass="mypassword",NameOrIds="names")
}
}
