% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab2.R
\name{rampFastMetaFromPath}
\alias{rampFastMetaFromPath}
\title{Use fast search algorithm to find all pathways from given analytes}
\usage{
rampFastMetaFromPath(pathway)
}
\arguments{
\item{pathway}{a string or a data.fram that contains all pathways}
}
\value{
a data.frame that contains all search results
}
\description{
Use fast search algorithm to find all pathways from given analytes
}
