% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab2.R
\name{rampFastMetaFromPath_InputFile}
\alias{rampFastMetaFromPath_InputFile}
\title{Generate data.frame that has all search results from given files}
\usage{
rampFastMetaFromPath_InputFile(infile, conpass = NULL, dbname = "ramp",
  username = "root", host = "localhost")
}
\arguments{
\item{infile}{a file object given from files}

\item{conpass}{password for database access (string)}

\item{dbname}{name of the mysql database (default is "ramp")}

\item{username}{username for database access (default is "root")}
}
\value{
a data.frame either from multiple csv file
or search through by a txt file.
}
\description{
identifing the file type, then it returns table output to 
shiny renderTable function as preview of searching data
}
