% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab1.R
\name{rampFastMetaInPath}
\alias{rampFastMetaInPath}
\title{Use fast algorithm to search through database to find analytes}
\usage{
rampFastMetaInPath(synonym, options = TRUE)
}
\arguments{
\item{synonym}{a string,or a data.frame that has all analytes' name user
want to search for}

\item{options}{a logic value used to define output (deprecate)}
}
\value{
a data.frame that contains all search results
}
\description{
Use fast algorithm to search through database to find analytes
}
