% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab1.R
\name{rampFastMetaInPath2}
\alias{rampFastMetaInPath2}
\title{Query: given a synonym, returns other genes or metabolites from the same pathway}
\usage{
rampFastMetaInPath2(synonym)
}
\arguments{
\item{synonym}{synonym (character string)}
}
\description{
Query: given a synonym, returns other genes or metabolites from the same pathway
}
