% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab4.R
\name{rampFastOneCata}
\alias{rampFastOneCata}
\title{Return analytes that has catalyzation relation with given analyte}
\usage{
rampFastOneCata(synonym, con)
}
\arguments{
\item{synonym}{string or data.frame contains given analyte}

\item{con}{a connection object returned from the function connectToRaMP()}
}
\value{
a data.frame that contains search results
}
\description{
Return analytes that has catalyzation relation with given analyte
}
