% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampMultipleInput.R
\name{rampFastPathFromMeta}
\alias{rampFastPathFromMeta}
\title{Search pathways from list of analytes in fast algorithm}
\usage{
rampFastPathFromMeta(synonym)
}
\arguments{
\item{synonym}{a string or data.frame that contains given analytes}
}
\value{
a data.frame that contains search results
}
\description{
Search pathways from list of analytes in fast algorithm
}
