% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{rampFastPathFromSource}
\alias{rampFastPathFromSource}
\title{Fast search given a list of metabolites source Id}
\usage{
rampFastPathFromSource(sourceid, find_synonym = FALSE, db = RaMP())
}
\arguments{
\item{sourceid}{a vector of synonym that need to be searched}

\item{find_synonym}{bool if find all synonyms or just return same synonym}

\item{db}{a RaMP database object}
}
\value{
a list contains all metabolits as name and pathway inside.
}
\description{
Fast search given a list of metabolites source Id
}
