% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab4.R
\name{rampFileOfAnalytes_tab4}
\alias{rampFileOfAnalytes_tab4}
\title{Generate data.frame from given files}
\usage{
rampFileOfAnalytes_tab4(infile)
}
\arguments{
\item{infile}{a file object given from files}
}
\value{
a data.frame either from multiple csv file
or search through by a txt file.
}
\description{
identifing the file type, then it returns table output to 
shiny renderTable function as preview of searching data
}
