% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab5.R
\name{rampFileOfBiofluid}
\alias{rampFileOfBiofluid}
\title{Generate search results based on given file}
\usage{
rampFileOfBiofluid(infile)
}
\arguments{
\item{infile}{a file object given by shiny}
}
\value{
search results of biofluid or analytes
}
\description{
Generate search results based on given file
}
