% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab2.R
\name{rampFileOfPathways_tab2}
\alias{rampFileOfPathways_tab2}
\title{Generate data.frame that has all search results from given files}
\usage{
rampFileOfPathways_tab2(infile)
}
\arguments{
\item{infile}{a file object given from files}
}
\value{
a data.frame either from multiple csv file
or search through by a txt file.
}
\description{
identifing the file type, then it returns table output to 
shiny renderTable function as preview of searching data
}
