% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{rampFindClassInfoFromSourceId}
\alias{rampFindClassInfoFromSourceId}
\title{Get class info for an input of metabolite source Ids}
\usage{
rampFindClassInfoFromSourceId(sourceIds)
}
\arguments{
\item{sourceIds}{a vector of analytes (genes or metabolites) that need to be searched}
}
\value{
a dataframe of chemClass info
}
\description{
Get class info for an input of metabolite source Ids
}
