% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{rampFindSynonymFromSynonym}
\alias{rampFindSynonymFromSynonym}
\title{Find all synonym from a given metabolite's name
This function is used to filter out some super common synonyms like glyceride
Now, this function only format the user input, so the user vector, dataframe,
and entire string separated by comma are working.}
\usage{
rampFindSynonymFromSynonym(
  synonym,
  full = FALSE,
  find_synonym = FALSE,
  con = NULL
)
}
\arguments{
\item{synonym}{name to search for}

\item{full}{bool if return whole data.frame}

\item{find_synonym}{bool if find all synonyms or just return same synonym}

\item{con}{MySQL connection (MySQLConnection)}
}
\value{
a data frame that contains synonym in the first column rampId in the second column
}
\description{
Find all synonym from a given metabolite's name
This function is used to filter out some super common synonyms like glyceride
Now, this function only format the user input, so the user vector, dataframe,
and entire string separated by comma are working.
}
