% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab3.R
\name{rampFisherTest}
\alias{rampFisherTest}
\title{Fisher test for given list of pathways}
\usage{
rampFisherTest(pathway_meta_list, num_user_metabolites, FisherPathwayTable)
}
\arguments{
\item{num_user_metabolites}{number of metabolites given by user when they want to 
search for pathways.}

\item{pathway_list}{The list contains pathway as names and metabolites in that pathway 
under each names (list)}
}
\value{
a data.frame contains all fisher test result with pathway name
as column name
}
\description{
From user input, the function accept a list of pathways, and number of metabolites 
from which the pathways are
}
