% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab3.R
\name{rampFisherTestData}
\alias{rampFisherTestData}
\title{Generate raw data for fisher test based on the given output
Output is from rampFastMetaFromPathway
Required format}
\usage{
rampFisherTestData(rampOut, metaboliteOrGene)
}
\arguments{
\item{rampOut}{The data frame generated by rampFastPathFromMeta}

\item{metaboliteOrGene}{specify whether to do test on 'metabolite' or 'gene'}
}
\value{
a data frame with three columns pathwayID, Number in pathway,
number out of pathway
}
\description{
Generate raw data for fisher test based on the given output
Output is from rampFastMetaFromPathway
Required format
}
