% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab3.R
\name{rampGenerateBarPlot}
\alias{rampGenerateBarPlot}
\title{Return a list named by pathway}
\usage{
rampGenerateBarPlot(df)
}
\arguments{
\item{df}{A dataframe that has information for bar plot}
}
\description{
each pathway contains all metabolites inside that pathway
}
