% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralFunctions.R
\name{rampGenesFromComp}
\alias{rampGenesFromComp}
\title{Send query to databases to get metabolites}
\usage{
rampGenesFromComp(names, maxItems, geneOrcompound = NULL, conpass = NULL,
  dbname = "ramp", username = "root")
}
\arguments{
\item{names}{the string that user defined for given synonym of metablite}

\item{maxItems}{the integer given from slider bar to limit maximum elements
returned by query}

\item{geneOrcompound}{string defined from radio input that describes type
of returned metablites from query.}

\item{conpass}{password for database access (string)}

\item{dbname}{name of the mysql database (default is "ramp")}

\item{username}{username for database access (default is "root")}
}
\value{
a dataframe that contain given synonym as column name
}
\description{
From user supplied synonym metabolites, it search through whole databases 
, and returns all metabolites which are in the same pathways.
}
\examples{
\dontrun{
ramp <- rampGenesFromComp("VitamineE",20,"compound",conpass="mysqlpassword")
}
}
