% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab3.R
\name{rampHcOutput}
\alias{rampHcOutput}
\title{highchart output for RaMP and fisher test.}
\usage{
rampHcOutput(x_data, y_data, type = "column", event_func)
}
\arguments{
\item{x_data}{vector contains data for categorical x-axis}

\item{y_data}{vector contains frequency of each pathway}

\item{type}{plot's type of this highcharter object}

\item{event_func}{Javascript code that define the click event}
}
\value{
highcharter object
}
\description{
Based on given x,y data, type and click event, it returns a highcharter object
to highchartOutput function to display bar plot.
}
