% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampDumpImport.R
\name{rampImportDumpFile}
\alias{rampImportDumpFile}
\title{Import RaMP database automatically from dump file that is coming with 
the Package}
\usage{
rampImportDumpFile(dbname, password)
}
\arguments{
\item{dbname}{The database name user defined}

\item{password}{The password for MySQL server}
}
\description{
Import RaMP database automatically from dump file that is coming with 
the Package
}
