% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralFunctions.R
\name{rampNameFromPath}
\alias{rampNameFromPath}
\title{Send query to databases to get metabolites}
\usage{
rampNameFromPath(names, maxItems, geneOrCompound, con)
}
\arguments{
\item{names}{the string value that describes pathway name.}

\item{maxItems}{the integer value given to define maximum items
return by this query.}

\item{geneOrCompound}{the string value that define types (gene or
compound) of returned metabolites.}

\item{con}{a connection object returned from the function connectToRaMP()}
}
\value{
a dataframe that has given pathwayname as column names
ramp <- rampNameFromPath("Glucose-Alanine Cycle",20,compound)
}
\description{
From given pathway name, it searches through whole databases to find
all metabolites in that pathway.
}
