% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampMultipleInput.R
\name{rampOneFisherTest}
\alias{rampOneFisherTest}
\title{Generate Result of fisher.test for only one column on highcharter}
\usage{
rampOneFisherTest(pathway_list, pathway, num_of_meta)
}
\arguments{
\item{pathway_list}{a list that has pathways as name,metabolites in 
sublist}

\item{pathway}{the string that stand for a pathway}

\item{num_of_meta}{an int that describe how many analytes in this pathway}
}
\value{
a dataframe that demonstrates stats result, p value and method
}
\description{
this function will be deprecated in next version
}
