% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralFunctions.R
\name{rampOntoOut}
\alias{rampOntoOut}
\title{The function send query to databases to find ontology or metabolites}
\usage{
rampOntoOut(synonym, maxItems)
}
\arguments{
\item{synonym}{string value that describe given metabolites or ontology}

\item{maxItems}{integer value that describe maximum items returned by query}
}
\value{
a dataframe that has given synonym as column name.
}
\description{
Based on ontology relationship, it search throught databases to find 
metabolites or ontology from given name which could be either ontology 
location or metabolite's synonym.
}
