% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralFunctions.R
\name{rampPathFromMeta}
\alias{rampPathFromMeta}
\title{Send query to database that find pathway}
\usage{
rampPathFromMeta(synonym, maxItems, conpass, dbname = "ramp",
  username = "root")
}
\arguments{
\item{synonym}{the string that describe given synonym name.}

\item{maxItems}{the integer that limit maximum returned by}

\item{conpass}{password for database access (string)}

\item{dbname}{name of the mysql database (default is "ramp")}

\item{username}{username for database access (default is "root")
query}
}
\value{
a dataframe that has given synonym as column name
}
\description{
From given metabolite's synonym, it search through all databases to find
all pathways that have the metabolite.
}
