% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralFunctions.R
\name{rampTablize}
\alias{rampTablize}
\title{Generate HTML table output from given data frame}
\usage{
rampTablize(data, num_row = 5)
}
\arguments{
\item{data}{The data converted to table format (data.frame)}

\item{num_row}{The rows of table converted from raw data (integer)}
}
\value{
A string that contains all html output
}
\description{
The HTML output is accepted by shiny HTML output, and 
form well-format table on webpage
}
