% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaMP.R
\name{removeLocalRampDB}
\alias{removeLocalRampDB}
\title{Remove a RaMP database version from local file cache}
\usage{
removeLocalRampDB(version = "none")
}
\arguments{
\item{version}{a ramp version as a string argument, e.g. '2.4.0'. The version parameter can be set to 'all'.}
}
\description{
Remove a RaMP database version from local file cache
}
