% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRaMP.R
\name{run_plumber_app}
\alias{run_plumber_app}
\title{Run plumber API}
\usage{
run_plumber_app(api_host = "127.0.0.1", api_port = 5762)
}
\arguments{
\item{api_host}{host}

\item{api_port}{port}
}
\description{
Run plumber API
}
