% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnPathwaysEnrich_InputAnalytes.R
\name{runCombinedFisherTest}
\alias{runCombinedFisherTest}
\title{Do fisher test for only one pathway from search result
clicked on highchart}
\usage{
runCombinedFisherTest(
  analytes,
  background = NULL,
  biospecimen_background = NULL,
  total_genes = 20000,
  min_analyte = 2,
  MCall = T,
  alternative = "less"
)
}
\arguments{
\item{analytes}{a vector of analytes (genes or metabolites) that need to be searched}

\item{background}{optional vector of all metabolites detected in study. This will be used as the background for the Fisher's contingency table for metabolites. If left NULL, all metabolites in RaMP originating in the pathway database of origin are used as background for testing}

\item{biospecimen_background}{If NULL, test all metabolites in RaMP or custom panel as Fisher's background. Else, use background for specific biospecimens. Choices are "Blood", "Adipose", "Heart", "Urine", "Brain", "Liver", "Kidney", "Saliva", and "Feces"}

\item{total_genes}{number of genes analyzed in the experiment (e.g. background) (default is 20000, with assumption that analyte_type is "genes")}

\item{min_analyte}{if the number of analytes (gene or metabolite) in a pathway is
< min_analyte, do not report}

\item{MCall}{T/F if true, all pathways are used for multiple comparison corrections; if false, only pathways covering user analytes will be used (default is "T")}

\item{alternative}{alternative hypothesis test passed on to fisher.test().  Options are two.sided, greater, or less (default is "less")}

\item{NameOrIds}{whether input is "names" or "ids" (default is "ids", must be the same for analytes and background)}
}
\value{
a list containing two entries: [[1]] fishresults, a dataframe containing pathways with Fisher's p values (raw and with FDR and Holm adjustment), number of user analytes in pathway, total number of analytes in pathway, and pathway source ID/database. [[2]] analyte_type, a string specifying the type of analyte input into the function ("genes", "metabolites", or "both")
}
\description{
Do fisher test for only one pathway from search result
clicked on highchart
}
\examples{
\dontrun{
pkg.globals <- setConnectionToRaMP(dbname = "ramp2", username = "root", conpass = "", host = "localhost")
pathwaydf <- getPathwayFromAnalyte(c("MDM2", "TP53", "glutamate", "creatinine"),
  NameOrIds = "names"
)
fisher.results <- runCombinedFisherTest(pathwaydf = pathwaydf)
}
}
