% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnPathwaysEnrich_InputAnalytes.R
\name{runCustomFisherTest}
\alias{runCustomFisherTest}
\title{Do fisher test with user-supplied custom pathway definitions}
\usage{
runCustomFisherTest(
  analytes,
  background = "kegg",
  background_type = "database",
  pathways,
  alternative = "less"
)
}
\arguments{
\item{analytes}{a vector of analytes (genes or metabolites) that need to be searched}

\item{background}{background to be used for Fisher's tests.  If parameter 'background_type="database"', this parameter
is ignored (default="database"); if parameter 'background_type= "file"', then 'background' should be a file name (with
directory); if 'background_type="list"', then 'background' should be a vector of RaMP IDs}

\item{background_type}{type of background that is input by the user.  Options are "database" if user wants all
analytes from the RaMP database will be used; "file", if user wants to input a file with a list of background
analytes; "list", if user wants to input a vector of analyte IDs}

\item{pathways}{(String) Path to gmx file containing custom pathway definitions. GMX files are a tab-separated format that contain one analyte set per column, with the name of the set in the first row, and constituent analytes in subsequent rows}

\item{alternative}{alternative hypothesis test passed on to fisher.test().  Options are two.sided, greater, or less (default is "less")}
}
\value{
a dataframe with columns containing pathway ID, fisher's p value, user analytes in pathway, and total analytes in pathway
}
\description{
Do fisher test with user-supplied custom pathway definitions
}
