% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OntologyQueryFunctions.R
\name{runEnrichOntologies}
\alias{runEnrichOntologies}
\title{Enrichment analysis for metabolite ontology mappings}
\usage{
runEnrichOntologies(
  mets,
  namesOrIds = "ids",
  alternative = "less",
  minMets = 2,
  minOntologySize = 5,
  maxOntologySize = 1500,
  includeRaMPids = FALSE,
  backgroundType = "database",
  background = "database",
  db = RaMP()
)
}
\arguments{
\item{mets}{a vector of metabolites that need to be searched}

\item{namesOrIds}{whether input is "names" or "ids" (default is "ids", must be the same for mets and background)}

\item{alternative}{alternative hypothesis test passed on to fisher.test().  Options are two.sided, greater, or less (default is "less")}

\item{minMets}{minimum number of mets per pathway (pathways with < minMets mets are filtered out).}

\item{minOntologySize}{the minimum number of ontology members (genes and metabolites) to include the ontology in the output (default = 5)}

\item{maxOntologySize}{the maximum number of ontology members (genes and metabolites) to include the ontology in the output (default = 150)}

\item{includeRaMPids}{whether or not to include RaMP IDs in the output (TRUE/FALSE)}

\item{backgroundType}{type of background that is input by the user.  Options are "database" if user wants all
mets from the RaMP database will be used; "file", if user wants to input a file with a list of background
mets; "list", if user wants to input a vector of analyte IDs; "biospecimen", if user wants to specify a
biospecimen type (e.g. blood, adipose tissue, etc.) and have those biospecimen-specific mets used.  For genes,
only the "database" option is used.}

\item{background}{background to be used for Fisher's tests.  If parameter 'backgroundType="database"', this parameter
is ignored (default="database"); if parameter 'backgroundType= "file"', then 'background' should be a file name (with
directory); if 'backgroundType="list"', then 'background' should be a vector of RaMP IDs; if 'backgroundType="biospecimen"'
then users should specify one of the following: "Blood", "Adipose tissue", "Heart", "Urine", "Brain", "Liver", "Kidney",
"Saliva", and "Feces"}

\item{db}{a RaMP database object}
}
\value{
a dataframe with columns containing pathway ID, fisher's p value, user mets in pathway, and total mets in pathway
}
\description{
Enrichment analysis for metabolite ontology mappings
}
\examples{
\dontrun{
ontologies.enriched <- runEnrichOntologies(mets = c("hmdb:HMDB0000033","hmdb:HMDB0000052","hmdb:HMDB0000094", "hmdb:HMDB0000161","hmdb:HMDB0000168","hmdb:HMDB0000191","hmdb:HMDB0000201","chemspider:10026", "hmdb:HMDB0006059", "Chemspider:6405", "CAS:5657-19-2","hmdb:HMDB0002511", "chemspider:20171375", "CAS:133-32-4","CAS:5746-90-7", "CAS:477251-67-5", "hmdb:HMDB0000695", "chebi:15934", "CAS:838-07-3", "hmdb:HMDBP00789", "hmdb:HMDBP00283", "hmdb:HMDBP00284", "hmdb:HMDBP00850"))
}
}
