% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampReactionQueries.R
\name{runEnrichReactionClass}
\alias{runEnrichReactionClass}
\title{Enrichment analysis for analyte-reaction class mappings}
\usage{
runEnrichReactionClass(
  analytes,
  humanProtein = TRUE,
  alternative = "less",
  includeRaMPids = FALSE,
  db = RaMP()
)
}
\arguments{
\item{analytes}{a vector of analyte ids (genes or metabolites) that need to be searched. ID types accepted: chebi and uniprot}

\item{humanProtein}{require reactions to have a human protein (enzyme or transporter), default True}

\item{alternative}{alternative hypothesis test passed on to fisher.test(). Options are two.sided, greater, or less (default is "less")}

\item{includeRaMPids}{include internal RaMP identifiers (default is "FALSE")}

\item{db}{a RaMP database object}
}
\value{
returns a list of [[1]] EC_Level1Stats, dataframe with columns containing reaction class ID, fisher's p value, user analytes in reaction class, and total analytes in reaction class at the enzyme class level 1 , [[2]] EC_Level2Stats, dataframe with columns containing reaction class ID, fisher's p value, user analytes in reaction class, and total analytes in reaction class at the enzyme class level 2, [[3]] analyteType,  a string specifying the type of analyte input into the function ("genes", "metabolites", or "both"), and [[4]] result_type, a string specifying enrichment of reaction classes was performed
}
\description{
Enrichment analysis for analyte-reaction class mappings
}
