% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbWorker.R
\name{runQuery}
\alias{runQuery}
\title{Get results from an SQL query to a RaMP database}
\usage{
runQuery(sql, db = RaMP())
}
\arguments{
\item{sql}{`character(1)` with the SQL query to run.}

\item{db}{[RaMP()] object representing a RaMP database. By default
(`db = RaMP()`) a connection to the most recent version is established,
which will be downloaded first if it does not already exist in the local
cache.}
}
\value{
The result from the query.
}
\description{
Utility function to execute the provided SQL query and get its results from
a RaMP-DB database.
}
