% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampReactionQueries.R
\name{runReactionClassTest}
\alias{runReactionClassTest}
\title{runReactionClassTest Enrichment analysis for analyte-reaction class mappings}
\usage{
runReactionClassTest(
  analytes,
  humanProtein = TRUE,
  alternative = "less",
  includeRaMPids = FALSE,
  db = RaMP()
)
}
\arguments{
\item{analytes}{a vector of analyte ids (genes or metabolites) that need to be searched. ID types accepted: chebi and uniprot}

\item{humanProtein}{require reactions to have a human protein (enzyme or transporter), default True}

\item{alternative}{alternative hypothesis test passed on to fisher.test(). Options are two.sided, greater, or less (default is "less")}

\item{includeRaMPids}{include internal RaMP identifiers (default is "FALSE")}

\item{db}{a RaMP database object}
}
\value{
returns a 2 dataframes with columns containing reaction class ID, fisher's p value, user analytes in reaction class, and total analytes in reaction class
}
\description{
runReactionClassTest Enrichment analysis for analyte-reaction class mappings
}
