% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{segregateDataBySource}
\alias{segregateDataBySource}
\title{Separate input ids into lists based on database of origin}
\usage{
segregateDataBySource(input_RampIds)
}
\arguments{
\item{input_RampIds}{list of analyte ramp IDs}
}
\value{
pathway lists separated by source db
}
\description{
Separate input ids into lists based on database of origin
}
\author{
Andrew Christopher Patt
}
