% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processNewRamp.R
\name{sysdataObject}
\alias{sysdataObject}
\title{sysdataObject() generates sysdata.rda object from exdata dir (genes_overlap_matrix.RData,metabolites_overlap_matrix.RData,analytes_overlap_matrix.RData,FT_data.Rdata)
the output is availble only for RaMP internal functions (runFisherTest() & findCluster())}
\usage{
sysdataObject()
}
\value{
sysdata.rda object in R directory with 3 Overlap Matrices (genes_result, metabolites_result, analyte_result) & dataframes generated from processData function.
}
\description{
sysdataObject() generates sysdata.rda object from exdata dir (genes_overlap_matrix.RData,metabolites_overlap_matrix.RData,analytes_overlap_matrix.RData,FT_data.Rdata)
the output is availble only for RaMP internal functions (runFisherTest() & findCluster())
}
