% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralFunctions.R
\name{updateOverlapMatrices}
\alias{updateOverlapMatrices}
\title{Update and save the overlap matrices based on current version of RaMP}
\usage{
updateOverlapMatrices(method, all, conpass, host = "localhost",
  dbname = "ramp", username = "root")
}
\arguments{
\item{method}{a string that specifies algorithm to compute overlap matrix
should be 'balanced' or 'weighted'}

\item{all}{a string that specifies which matrices to compute, should be in
'all','analyte'}

\item{conpass}{a string that specifies password for database connection}

\item{host}{a string that specifes host for database connection}

\item{dbname}{a string that specifies database name of MySQL database}

\item{username}{a string that specifies name of MySQL database}
}
\description{
Update and save the overlap matrices based on current version of RaMP
}
