% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapMatrix.R
\name{updateOverlapMatrix}
\alias{updateOverlapMatrix}
\title{Update the overlap matrix for store in the shiny app directory}
\usage{
updateOverlapMatrix(min_analyte, method, together, conpass,
  host = "localhost", dbname = "ramp", username = "root")
}
\arguments{
\item{min_analyte}{a int that specifies the minimum of analytes the
pathway should have to be considered compute for overlap matrix}

\item{method}{a string that specifies the way to compute overlap matrix,
must be 'balanced' or 'weighted'}

\item{together}{a boolean value to compute overlap matrix for
gene/metabolites separatly or together}

\item{conpass}{a string that specifies password for database connection}

\item{host}{a string that specifes host for database connection}

\item{dbname}{a string that specifies database name of MySQL database}

\item{username}{a string that specifies name of MySQL database}
}
\description{
Update the overlap matrix for store in the shiny app directory
}
