% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writingFunctions.R
\name{write_FishersResults}
\alias{write_FishersResults}
\title{Function that writes Fishers Test results, after clustering to a CSV file}
\usage{
write_FishersResults(fishResults = "none", outputFile = "none", rampId = FALSE)
}
\arguments{
\item{fishResults}{a data frame returned by function runCombinedFisherTest()}

\item{outputFile}{name of output file}

\item{rampId}{whether or not to include rampId (default is FALSE)}
}
\description{
Function that writes Fishers Test results, after clustering to a CSV file
}
