% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writingFunctions.R
\name{write_to_csv}
\alias{write_to_csv}
\title{The function write a data.frame to a csv files.}
\usage{
write_to_csv(df, f.name)
}
\arguments{
\item{df}{a data frame returned by functions that queires database.}

\item{f.name}{a string that represents output file name.}
}
\description{
The function write a data.frame to a csv files.
}
