% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writingFunctions.R
\name{writeFishersResults}
\alias{writeFishersResults}
\title{Function that writes Fishers Test results, after clustering to a CSV file}
\usage{
writeFishersResults(
  fishResults = "none",
  outputFile = "none",
  includeRaMPids = FALSE
)
}
\arguments{
\item{fishResults}{a data frame returned by function runCombinedFisherTest()}

\item{outputFile}{name of output file}

\item{includeRaMPids}{whether or not to include rampIds (default is FALSE)}
}
\description{
Function that writes Fishers Test results, after clustering to a CSV file
}
