% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writingFunctions.R
\name{writePathwaysToCSV}
\alias{writePathwaysToCSV}
\title{Function that writes output from getPathwayFromAnalyte() to a CSV file}
\usage{
writePathwaysToCSV(mypathways = "none", outputfile = "none")
}
\arguments{
\item{mypathways}{data frame returned by function getPathwayFromAnalyte()}

\item{outputfile}{name of output file}
}
\description{
Function that writes output from getPathwayFromAnalyte() to a CSV file
}
