% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab3.R
\name{WY_pathway_adjustment}
\alias{WY_pathway_adjustment}
\title{Fisher's test with the p value permutation approach developed by Westfall and Young to account for overlapping pathway annotations}
\usage{
WY_pathway_adjustment(
  pathwaydf,
  backgrounddf,
  total_metabolites = NULL,
  total_genes = 20000,
  analyte_type = "metabolites",
  conpass = NULL,
  dbname = "ramp",
  username = "root",
  host = "localhost",
  socket = NULL,
  repetitions = 100,
  metaboliteClusters = NULL,
  ncpus = 1
)
}
\arguments{
\item{pathwaydf}{a data frame resulting from getPathwayFromAnalyte}

\item{backgrounddf}{optional dataframe resulting from getPathwayFromAnalyte run on all analytes detected in study. This will be used as the background for the Fisher's contingency table. If left NULL, all analytes in RaMP are used as background}

\item{total_metabolites}{number of metabolites analyzed in the experiment (e.g. background) (default is 1000; set to 'NULL' to retrieve total number of metabolites that map to any pathway in RaMP). Assumption that analyte_type is "metabolite")}

\item{total_genes}{number of genes analyzed in the experiment (e.g. background) (default is 20000, with assumption that analyte_type is "genes")}

\item{analyte_type}{"metabolites" or "genes" (default is "metabolites")}

\item{conpass}{password for database access (string)}

\item{dbname}{name of the mysql database (default is "ramp")}

\item{username}{username for database access (default is "root")}

\item{host}{host name for database access (default is "localhost")}

\item{socket}{(optional) location of mysql.sock file}

\item{repetitions}{number of times p values are permuted}

\item{metaboliteClusters}{Optional output of cluster_metabolites}

\item{ncpus}{optional parameter indicating number of cpus available for parallel computation}
}
\value{
a dataframe with columns containing pathway ID, fisher's p value, user analytes in pathway, total analytes in pathway, and Westfall-Young permuted p-values
}
\description{
Fisher's test with the p value permutation approach developed by Westfall and Young to account for overlapping pathway annotations
}
