% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{buildFrequencyTables}
\alias{buildFrequencyTables}
\title{Return all analytes that map to unique pathways in a pathwaydf}
\usage{
buildFrequencyTables(db = RaMP(), inputdf, pathway_definitions, analyte_type)
}
\arguments{
\item{db}{a RaMP database object}

\item{inputdf}{internal df with pathwayramp ids}

\item{pathway_definitions}{If "RaMP" (default), use pathway definitions within RaMP-DB. Else, supply path to gmx file containing custom pathway definitions. GMX files are a tab-separated format that contain one analyte set per column, with the name of the set in the first row, and constituent analytes in subsequent rows}

\item{analyte_type}{"genes" or "metabolites"}
}
\value{
dataframe of all analytes that map to the input pathways
}
\description{
Return all analytes that map to unique pathways in a pathwaydf
}
\author{
Andrew Christopher Patt
}
