% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampChemClassQueries.R
\name{chemicalClassEnrichment}
\alias{chemicalClassEnrichment}
\title{returns chemical class information comparing a metabolite subset to a metabolite population}
\usage{
chemicalClassEnrichment(classData)
}
\arguments{
\item{classData}{a chemical class result object from chemicalClassSurvey}
}
\value{
a data frame containing chemical class enrichment statistics
}
\description{
returns chemical class information comparing a metabolite subset to a metabolite population
}
