% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampChemClassQueries.R
\name{chemicalClassEnrichment}
\alias{chemicalClassEnrichment}
\title{returns chemical class information comparing a metabolite subset to a metabolite population}
\usage{
chemicalClassEnrichment(classData)
}
\arguments{
\item{classData}{a chemical class result object from chemicalClassSurvey}
}
\value{
a data frame containing chemical class enrichment statistics
}
\description{
returns chemical class information comparing a metabolite subset to a metabolite population
}
\examples{
\dontrun{
# metabolite list of interest
metList = c('hmdb:HMDB0000056',
            'hmdb:HMDB0000439',
            'hmdb:HMDB0000479',
            'hmdb:HMDB0000532',
            'hmdb:HMDB0001015',
            'hmdb:HMDB0001138',
            'hmdb:HMDB0029159',
            'hmdb:HMDB0029412',
            'hmdb:HMDB0034365',
            'hmdb:HMDB0035227',
            'hmdb:HMDB0007973',
            'hmdb:HMDB0008057',
            'hmdb:HMDB0011211')

# the background population can be a separate ID list (preferred) or all database entries 
# (skip pop parameter).
metClassResult <- chemicalClassSurvey(mets = mets, conpass, dbname, host, username)

enrichedClassStats <- chemicalClassEnrichment(metClassResult)
}
}
