% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectToRaMP.R
\name{connectToRaMP}
\alias{connectToRaMP}
\title{Connect to RaMP database (requires mysql password when running locally)}
\usage{
connectToRaMP(dbname = "ramp", username = "root", password = NULL,
  host = NULL)
}
\arguments{
\item{dbname}{the name of the database (by default a string that is database name including all tables}

\item{username}{a string that is username for database (default: root)}

\item{password}{a string this is password for database}

\item{host}{a string that stand for host}
}
\description{
Connect to RaMP database (requires mysql password when running locally)
}
\examples{
\dontrun{
con <- connectToRaMP(dbname="ramp",username="root",password="mypassword")
}
}
