% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{filterEnrichResults}
\alias{filterEnrichResults}
\title{Filter enrichment results by p-value cutoff for display and clustering}
\usage{
filterEnrichResults(enrichResults, pValType = "fdr", pValCutoff = 0.1)
}
\arguments{
\item{enrichResults}{The data frame generated by runEnrich...}

\item{pValType}{Specifies which p-value to use as the filter threshold.
Permitted values are 'pval' for raw p-value and 'fdr' or 'holm' p-value corrections. Default is 'fdr'.}

\item{pValCutoff}{return results where pValType p-values are < pValCutoff}
}
\value{
enrichResults with the fishresults dataframe filtered based on thresholds set
}
\description{
Filter enrichment results by p-value cutoff for display and clustering
}
\examples{
\dontrun{
analyteList <- c("MDM2", "TP53", "glutamate", "creatinine")

fisher.results <- runEnrichPathways(analytes = analyteList, namesOrIds = 'names')
filtered.fisher.results <- filterEnrichResults(fisher.results, pValType='fdr', pValCutoff = 0.10)
}
}
