% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab3.R
\name{FilterFishersResults}
\alias{FilterFishersResults}
\title{Filter pathways by p-value cutoff for display and clustering}
\usage{
FilterFishersResults(fishers_df, p_holmadj_cutoff = NULL,
  p_fdradj_cutoff = NULL)
}
\arguments{
\item{fishers_df}{The data frame generated by runFisherTest}

\item{p_holmadj_cutoff}{return pathways where Holm adjusted pvalues are < p_holmadj_cutoff}

\item{p_fdradj_cutoff}{return pathways where FDR adjusted pvalues are < p_fdradj_cutoff}
}
\value{
dataframe of fishers results with only significant pathways
}
\description{
Filter pathways by p-value cutoff for display and clustering
}
\examples{
\dontrun{
pathwaydf<-getPathwayFromAnalyte(c("MDM2","TP53","glutamate","creatinine"),
                NameOrIds="names", conpass=conpass)
fisher.results <- runCombinedFisherTest(pathwaydf=pathwaydf,conpass=conpass)
filtered.fisher.results <- FilterFishersResults(fisher.results,p_holmadj_cutoff=0.05)
}
}
