% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{FilterFishersResults}
\alias{FilterFishersResults}
\title{Filter pathways by p-value cutoff for display and clustering}
\usage{
FilterFishersResults(fishers_df, pval_type = "fdr", pval_cutoff = 0.1)
}
\arguments{
\item{fishers_df}{The data frame generated by runFisherTest}

\item{pval_type}{Specifies which p-value to use as the filter threshold.
Permitted values are 'pval' and 'fdr' for chemical class and pathway enrichment.
Pathway enrichment also includes an optional 'holm' value for holm p-value corrections. Default is 'fdr'.}

\item{pval_cutoff}{return pathways where pval_type p-values are < pval_cutoff}
}
\value{
list:[[1]]Dataframe with pathway enrichment results, only significant pathways
[[2]]analyte type
}
\description{
Filter pathways by p-value cutoff for display and clustering
}
\examples{
\dontrun{
analyteList <- c("MDM2", "TP53", "glutamate", "creatinine")

fisher.results <- runCombinedFisherTest(analytes = analyteList, namesOrIds = 'names')
filtered.fisher.results <- FilterFishersResults(fisher.results, pval_type='fdr', pval_cutoff = 0.10)
}
}
