% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnPathwaysEnrich_InputAnalytes.R
\name{findCluster}
\alias{findCluster}
\title{Perform fuzzy multiple linkage partitioning clustering on pathways
identified by Fisher's test}
\usage{
findCluster(
  fishersDf,
  percAnalyteOverlap = 0.5,
  minPathwayToCluster = 2,
  percPathwayOverlap = 0.5,
  db = RaMP()
)
}
\arguments{
\item{fishersDf}{The full result object generated by runCombinedFisherTest}

\item{percAnalyteOverlap}{Minimum overlap for pathways to be considered similar
(Default = 0.5)}

\item{minPathwayToCluster}{Minimum number of 'similar' pathways required to start
a cluster (medoid) (Default = 3)}

\item{percPathwayOverlap}{Minimum overlap for clusters to merge (Default = 0.5)}

\item{db}{a RaMP database object}
}
\value{
list:[[1]] Pathway enrichment result with dataframe having a cluster assignment column added
[[2]] analyte type
[[3]] cluster assignment in the list form
}
\description{
Perform fuzzy multiple linkage partitioning clustering on pathways
identified by Fisher's test
}
\examples{
\dontrun{
pathwaydf <- getPathwayFromAnalyte(c(
"ensembl:ENSG00000135679", "hmdb:HMDB0000064",
"hmdb:HMDB0000148", "ensembl:ENSG00000141510"
))

fisher.results <- runCombinedFisherTest(pathwaydf = pathwaydf)

clustered.fisher.results <- findCluster(fisher.results)
}
}
