% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab3.R
\name{findCluster}
\alias{findCluster}
\title{Perform fuzzy multiple linkage partitioning clustering on pathways identified by
Fisher's test}
\usage{
findCluster(
  fishers_df,
  perc_analyte_overlap = 0.5,
  min_pathway_tocluster = 2,
  perc_pathway_overlap = 0.5
)
}
\arguments{
\item{fishers_df}{The data frame generated by runFisherTest}

\item{perc_analyte_overlap}{Minimum overlap for pathways to be considered similar
(Default = 0.5)}

\item{min_pathway_tocluster}{Minimum number of 'similar' pathways required to start
a cluster (medoid) (Default = 3)}

\item{perc_pathway_overlap}{Minimum overlap for clusters to merge (Default = 0.5)}
}
\value{
list:[[1]] Pathway enrichment result dataframe with cluster assignment column added
[[2]] analyte type
[[3]] cluster assignment in the list form
}
\description{
Perform fuzzy multiple linkage partitioning clustering on pathways identified by
Fisher's test
}
\examples{
\dontrun{
pathwaydf<-getPathwayFromAnalyte(c("MDM2","TP53","glutamate","creatinine"),
                NameOrIds="names", conpass=conpass)
fisher.results <- runCombinedFisherTest(pathwaydf=pathwaydf,conpass=conpass)
filtered.fisher.results <- FilterFishersResults(fisher.results,p_holmadj_cutoff=0.05)
filteredclust.fisher.results <- findCluster(filtered.fisher.results)
}
}
