% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnAnalytes_InputPathways.R
\name{getAnalyteFromPathway}
\alias{getAnalyteFromPathway}
\title{Use fast search algorithm to find all pathways from given analytes}
\usage{
getAnalyteFromPathway(
  pathway,
  match = "exact",
  analyteType = "both",
  maxPathwaySize = Inf,
  namesOrIds = "names",
  db = RaMP()
)
}
\arguments{
\item{pathway}{a string or a vector of strings that contains pathways of interest}

\item{match}{type of matching to use, options are "exact" or "fuzzy".  The default is "exact".}

\item{analyteType}{a string denoting the type of analyte to return ("gene", "metabolite", "both")}

\item{maxPathwaySize}{(default Inf), trims returned results to pathways that have fewer than this number}

\item{namesOrIds}{are the input pathways input as pathway names or as pathway ids
of genes and metabolites}

\item{db}{a RaMP database object, if not specified a new one is created with RaMP::RaMP()}
}
\value{
a data.frame that contains all search results
}
\description{
Use fast search algorithm to find all pathways from given analytes
}
\examples{
\dontrun{

getAnalyteFromPathway( pathway=c("Wnt Signaling Pathway", "Sphingolipid Metabolism"), db = rampDB )

}
}
