% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnAnalytes_InputPathways.R
\name{getAnalyteFromPathway}
\alias{getAnalyteFromPathway}
\title{Use fast search algorithm to find all pathways from given analytes}
\usage{
getAnalyteFromPathway(pathway)
}
\arguments{
\item{pathway}{a string or a vector of strings that contains pathways of interest}
}
\value{
a data.frame that contains all search results
}
\description{
Use fast search algorithm to find all pathways from given analytes
}
\examples{
\dontrun{
# To query one pathway:
myanalytes <- getAnalyteFromPathway(pathway="sphingolipid metabolism")

# To query multiple pathways:
pkg.globals <- setConnectionToRaMP(dbname="ramp2",username="root",conpass="",host = "localhost")
myanalytes <- getAnalyteFromPathway(pathway=c("De Novo Triacylglycerol Biosynthesis", 
"sphingolipid metabolism"))
}
}
