% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnAnalytes_InputPathways.R
\name{getAnalyteFromPathway}
\alias{getAnalyteFromPathway}
\title{Use fast search algorithm to find all pathways from given analytes}
\usage{
getAnalyteFromPathway(
  pathway,
  match = "exact",
  analyte_type = "both",
  max_pathway_size = Inf,
  names_or_ids = "names"
)
}
\arguments{
\item{pathway}{a string or a vector of strings that contains pathways of interest}

\item{match}{type of matching to use, options are "exact" or "fuzzy".  The default is "exact".}

\item{analyte_type}{a string denoting the type of analyte to return ("gene", "metabolite", "both")}

\item{max_pathway_size}{(default Inf), trims returned results to pathways that have fewer than this number}

\item{names_or_ids}{are the input pathways input as pathway names or as pathway ids
of genes and metabolites}
}
\value{
a data.frame that contains all search results
}
\description{
Use fast search algorithm to find all pathways from given analytes
}
\examples{
\dontrun{
# To query one pathway:
myanalytes <- getAnalyteFromPathway2(pathway="sphingolipid metabolism")

# To query multiple pathways:
pkg.globals <- setConnectionToRaMP(dbname="ramp2",username="root",conpass="",host = "localhost")
myanalytes <- getAnalyteFromPathway2(pathway=c("De Novo Triacylglycerol Biosynthesis",
"sphingolipid metabolism"))
}
}
