% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab2.R
\name{getAnalyteFromPathway}
\alias{getAnalyteFromPathway}
\title{Use fast search algorithm to find all pathways from given analytes}
\usage{
getAnalyteFromPathway(
  pathway,
  conpass = NULL,
  dbname = "ramp",
  username = "root",
  host = "localhost"
)
}
\arguments{
\item{pathway}{a string or a vector of strings that contains pathways of interest}

\item{conpass}{password for database access (string)}

\item{dbname}{name of the mysql database (default is "ramp")}

\item{username}{username for database access (default is "root")}

\item{host}{host name for database access (default is "localhost")}
}
\value{
a data.frame that contains all search results
}
\description{
Use fast search algorithm to find all pathways from given analytes
}
\examples{
\dontrun{
con <- connectToRaMP(dbname="ramp",username="root",conpass="mypassword")
getAnalyteFromPathway <- pathway="sphingolipid metabolism",conpass="mypassword")
}
}
