% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metClassUtil.R
\name{getMetabClassTypes}
\alias{getMetabClassTypes}
\title{Returns class categories for metabolites}
\usage{
getMetabClassTypes(db = RaMP())
}
\value{
Returns metabolite class types
}
\description{
Returns class categories for metabolites
}
