% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OntologyQueryFunctions.R
\name{getMetaFromOnto}
\alias{getMetaFromOnto}
\title{function that query database to find analytes in given ontologies}
\usage{
getMetaFromOnto(ontology, db = RaMP())
}
\arguments{
\item{ontology}{a vector of ontology or ontologies delimited by new line character}

\item{db}{a RaMP database object}
}
\value{
dataframe that  contains searched analytes from given ontology
}
\description{
function that query database to find analytes in given ontologies
}
\examples{
\dontrun{
<<<<<<< HEAD
ontologies.of.interest <- c("Colon", "Liver", "Lung")

new.metabolites <- RaMP::getMetaFromOnto(db = rampDB, ontology = ontologies.of.interest)
=======
pkg.globals <- setConnectionToRaMP(
  dbname = "ramp2", username = "root",
  conpass = "", host = "localhost"
)
getMetaFromOnto("Adiposome")
>>>>>>> 12716f1d27e660ab70e18b100d2e9e2ff3da755a
}
}
