% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OntologyQueryFunctions.R
\name{getMetaFromOnto}
\alias{getMetaFromOnto}
\title{function that query database to find mets in given ontologies}
\usage{
getMetaFromOnto(ontology, db = RaMP())
}
\arguments{
\item{ontology}{a vector of ontology or ontologies delimited by new line character}

\item{db}{a RaMP database object}
}
\value{
dataframe that contains searched mets from given ontology
}
\description{
function that query database to find mets in given ontologies
}
\examples{
\dontrun{
ontologies.of.interest <- c("Colon", "Liver", "Lung")

new.metabolites <- RaMP::getMetaFromOnto(db = rampDB, ontology = ontologies.of.interest)
}
}
