% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OntologyQueryFunctions.R
\name{getMetaFromOnto}
\alias{getMetaFromOnto}
\title{function that query database to find analytes in given ontologies}
\usage{
getMetaFromOnto(db = RaMP(), ontology)
}
\arguments{
\item{ontology}{a vector of ontology or ontologies delimited by new line character}
}
\value{
dataframe that  contains searched analytes from given ontology
}
\description{
function that query database to find analytes in given ontologies
}
\examples{
\dontrun{
pkg.globals <- setConnectionToRaMP(
  dbname = "ramp2", username = "root",
  conpass = "", host = "localhost"
)
getMetaFromOnto("Adiposome")
}
}
