% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab5.R
\name{getMetaFromOnto}
\alias{getMetaFromOnto}
\title{function that query database to find analytes in given ontologies}
\usage{
getMetaFromOnto(
  ontology,
  conpass = NULL,
  dbname = "ramp",
  host = "localhost",
  username = "root"
)
}
\arguments{
\item{ontology}{a vector of ontology or ontologies delimited by new line character}

\item{conpass}{a database password}

\item{dbname}{a database name for the database connected}

\item{host}{host name for the database}

\item{username}{user name for the database}
}
\value{
dataframe that  contains searched analytes from given ontology
}
\description{
function that query database to find analytes in given ontologies
}
