% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OntologyQueryFunctions.R
\name{getOntoFromMeta}
\alias{getOntoFromMeta}
\title{Function that query database to find ontology information based on
the given list of analytes}
\usage{
getOntoFromMeta(analytes, namesOrIds = "ids", db = RaMP())
}
\arguments{
\item{analytes}{a vector of analytes or a analytes delimited by new line character}

\item{namesOrIds}{specify the type of given data}

\item{db}{a RaMP database object}
}
\value{
dataframe that contains searched ontology from given analytes
}
\description{
Function that query database to find ontology information based on
the given list of analytes
}
\examples{
\dontrun{
analytes.of.interest <- c("chebi:15422", "hmdb:HMDB0000064", "hmdb:HMDB0000148", "wikidata:Q426660")

new.ontologies <- RaMP::getOntoFromMeta(db = rampDB, analytes = analytes.of.interest)
}
}
