% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OntologyQueryFunctions.R
\name{getOntoFromMeta}
\alias{getOntoFromMeta}
\title{Function that query database to find ontology information based on
the given list of analytes}
\usage{
getOntoFromMeta(analytes, NameOrIds = "ids")
}
\arguments{
\item{analytes}{a vector of analytes or a analytes delimited by new line character}

\item{NameOrIds}{specify the type of given data}
}
\value{
dataframe that contains searched ontology from given analytes
}
\description{
Function that query database to find ontology information based on
the given list of analytes
}
\examples{
\dontrun{
pkg.globals <- setConnectionToRaMP(dbname = "ramp2", username = "root",
	conpass = "", host = "localhost")
getOntoFromMeta("hmdb:HMDB0071437")
}
}
