% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab5.R
\name{getOntoFromMeta}
\alias{getOntoFromMeta}
\title{Function that query database to find ontology information based on
the given list of analytes}
\usage{
getOntoFromMeta(
  analytes,
  conpass = NULL,
  dbname = "ramp",
  host = "localhost",
  username = "root",
  NameOrIds = "ids"
)
}
\arguments{
\item{analytes}{a vector of analytes or a analytes delimited by new line character}

\item{conpass}{password for the database}

\item{dbname}{database name for this database}

\item{host}{host for this database}

\item{username}{username for the database}

\item{NameOrIds}{specify the type of given data}
}
\value{
dataframe that contains searched ontology from given analytes
}
\description{
Function that query database to find ontology information based on
the given list of analytes
}
