% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnPathwaysEnrich_InputAnalytes.R
\name{getPathwayFromAnalyte}
\alias{getPathwayFromAnalyte}
\title{Function that search analytes (gene or compounds)  or a list of analytes and
returns associated pathways}
\usage{
getPathwayFromAnalyte(
  analytes = "none",
  findSynonym = FALSE,
  namesOrIds = "ids",
  includeRaMPids = FALSE,
  includeSmpdb = FALSE,
  minPathwaySize = 5,
  maxPathwaySize = 150,
  db = RaMP(),
  ...
)
}
\arguments{
\item{analytes}{a vector of analytes (genes or metabolites) that need to be searched}

\item{findSynonym}{find all synonyms or just return same synonym (T/F)}

\item{namesOrIds}{whether input is "names" or "ids" (default is "ids")}

\item{includeRaMPids}{include internal RaMP identifiers (default is "FALSE")}

\item{includeSmpdb}{Include pathways from smpdb/hmdb in analysis. Excluded by default since definitions are
highly redundant}

\item{minPathwaySize}{the minimum number of pathway members (genes and metabolites) to include the pathway in
the output (default = 5)}

\item{maxPathwaySize}{the maximum number of pathway memnbers (genes and metaboltes) to include the pathway in
the output (default = 150)}

\item{db}{a RaMP database object}

\item{...}{Internal Use - for handling deprecated parameter names}
}
\value{
a list contains all metabolites as name and pathway inside.
}
\description{
Function that search analytes (gene or compounds)  or a list of analytes and
returns associated pathways
}
\examples{
\dontrun{
getPathwayFromAnalyte(db = rampDB, namesOrIds="ids", c("ensembl:ENSG00000135679",
     "hmdb:HMDB0000064",
     "hmdb:HMDB0000148",
     "ensembl:ENSG00000141510"))

getPathwayFromAnalyte(db = rampDB, namesOrIds = "names", c("Serotonin", "Bilirubin", "Urea"))
}
}
