% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnPathwaysEnrich_InputAnalytes.R
\name{getPathwayFromAnalyte}
\alias{getPathwayFromAnalyte}
\title{Function that search analytes (gene or compounds)  or a list of analytes and
returns associated pathways}
\usage{
getPathwayFromAnalyte(analytes = NULL, find_synonym = FALSE, NameOrIds = "ids")
}
\arguments{
\item{analytes}{a vector of analytes (genes or metabolites) that need to be searched}

\item{find_synonym}{find all synonyms or just return same synonym (T/F)}

\item{NameOrIds}{whether input is "names" or "ids" (default is "ids")}
}
\value{
a list contains all metabolites as name and pathway inside.
}
\description{
Function that search analytes (gene or compounds)  or a list of analytes and
returns associated pathways
}
\examples{
\dontrun{
pkg.globals <- setConnectionToRaMP(dbname = "ramp2", username = "root", conpass = "", host = "localhost")
mypath <- getPathwayFromAnalyte(analytes = c("2-hydroxyglutarate", "glutamate"))
}
}
