% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab3.R
\name{getPathwayFromAnalyte}
\alias{getPathwayFromAnalyte}
\title{Function that search analytes (gene or compounds)  or a list of analytes and
returns associated pathways}
\usage{
getPathwayFromAnalyte(analytes = NULL, find_synonym = FALSE,
  conpass = NULL, host = "localhost", dbname = "ramp",
  username = "root", NameOrIds = "ids")
}
\arguments{
\item{analytes}{a vector of analytes (genes or metabolites) that need to be searched}

\item{find_synonym}{find all synonyms or just return same synonym (T/F)}

\item{conpass}{password for database access (string)}

\item{host}{host name for database access (default is "localhost")}

\item{dbname}{name of the mysql database (default is "ramp")}

\item{username}{username for database access (default is "root")}

\item{NameOrIds}{whether input is "names" or "ids" (default is "ids")}
}
\value{
a list contains all metabolites as name and pathway inside.
}
\description{
Function that search analytes (gene or compounds)  or a list of analytes and
returns associated pathways
}
\examples{
\dontrun{
mypath <- getPathwayFromAnalyte(analytes=c("2-hydroxyglutarate","glutamate"), conpass="mypassword")
}
}
