% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metClassUtil.R
\name{getPrefixesFromAnalytes}
\alias{getPrefixesFromAnalytes}
\title{Query to retrieve database ID prefixes for analyte types}
\usage{
getPrefixesFromAnalytes(db = RaMP(), analyteType = "gene")
}
\arguments{
\item{analyteType}{value to indicate the desired analyte type. Value one of 'gene' or 'metabolite'}
}
\value{
Returns list of database ID prefixes for selected 'gene' or 'metabolite'
}
\description{
Query to retrieve database ID prefixes for analyte types
}
