% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{getRaMPInfoFromAnalytes}
\alias{getRaMPInfoFromAnalytes}
\title{Internal function for extracting annotations, used by pathway and chemical enrichment test functions}
\usage{
getRaMPInfoFromAnalytes(
  db = RaMP(),
  analytes,
  NameOrIds = "ids",
  PathOrChem = "path"
)
}
\arguments{
\item{analytes}{a vector of analytes (genes or metabolites) that need to be searched}

\item{NameOrIds}{whether input is "names" or "ids" (default is "ids")}

\item{PathOrChem}{return "path" information for pathways or "chem" for chemical class}
}
\value{
a list of rampIds for "path" or a dataframe of chemClass info
}
\description{
Internal function for extracting annotations, used by pathway and chemical enrichment test functions
}
