% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampReactionQueries.R
\name{getReactionClassesForAnalytes}
\alias{getReactionClassesForAnalytes}
\title{getReactionClassesForAnalytes returns reactions class and EC numbers for a collection of input compound ids}
\usage{
getReactionClassesForAnalytes(
  db = RaMP(),
  analytes,
  multiRxnParticipantCount = 1,
  humanProtein = F
)
}
\arguments{
\item{db}{a RaMP databse object}

\item{analytes}{list of analyte ids}
}
\value{
returns a three dataframes of reaction EC classe information, one for each EC level
}
\description{
getReactionClassesForAnalytes returns reactions class and EC numbers for a collection of input compound ids
}
