% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampReactionQueries.R
\name{getReactionClassStatsOnAnalytes}
\alias{getReactionClassStatsOnAnalytes}
\title{Utility method that returns the number of rhea reaction count in each reaction class.}
\usage{
getReactionClassStatsOnAnalytes(humanProtein = T, db = RaMP())
}
\arguments{
\item{humanProtein}{boolean value indicating if reactions should be constrained to those having human proteins}

\item{db}{RaMP Object}
}
\description{
Utility method that returns the number of rhea reaction count in each reaction class.
}
