% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampReactionQueries.R
\name{getReactionsForRaMPGeneIds}
\alias{getReactionsForRaMPGeneIds}
\title{getReactionsForRaMPGeneIds returns reactions for a collection of ramp compound ids}
\usage{
getReactionsForRaMPGeneIds(
  rampGeneIds,
  onlyHumanMets = F,
  humanProtein = T,
  includeTransportRxns = F,
  rxnDirs = c("UN"),
  db = RaMP()
)
}
\arguments{
\item{rampGeneIds}{list of ramp compound ids}

\item{onlyHumanMets}{boolean to only return pathways containing only human metabolites (ChEBI ontology) (dev in progress)}

\item{humanProtein}{boolean to only control pathways catalyzed by a human proteins (having human Uniprot) (dev in progress)}

\item{includeTransportRxns}{if TRUE, returns metabolic and transport reactions}

\item{rxnDirs}{character vector of length > 1, specifying reaction directions to return  c("UN", "LR", "RL", "BD", "ALL"), default = c("UN")}

\item{db}{a RaMP database object}
}
\value{
returns a dataframe of reaction information for each ramp compound id
}
\description{
getReactionsForRaMPGeneIds returns reactions for a collection of ramp compound ids
}
