% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampReactionQueries.R
\name{getReactionsForSourceCompoundIds}
\alias{getReactionsForSourceCompoundIds}
\title{getReactionsForSourceCompoundIds returns reactions for a collection of ChEBI input Ids}
\usage{
getReactionsForSourceCompoundIds(
  compoundIds,
  onlyHumanMets = F,
  humanProtein = F,
  includeTransportRxns = F,
  rxnDirs = c("UN"),
  db = RaMP()
)
}
\arguments{
\item{compoundIds}{list of ChEBI compound ids (prefixed with chebi:)}

\item{onlyHumanMets}{boolean to only return pathways containing only human metabolites (ChEBI ontology) (dev in progress)}

\item{humanProtein}{boolean to only control pathways catalyzed by a human proteins (having human Uniprot) (dev in progress)}

\item{includeTransportRxns}{if TRUE, returns metabolic and transport reactions}

\item{rxnDirs}{character vector of length > 1, specifying reaction directions to return  c("UN", "LR", "RL", "BD", "ALL"), default = c("UN")}

\item{db}{a RaMP database object}
}
\value{
returns a dataframe of reaction information for each ramp compound id
}
\description{
getReactionsForSourceCompoundIds returns reactions for a collection of ChEBI input Ids
}
