% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotAnalyteOverlapPerRxnLevel}
\alias{plotAnalyteOverlapPerRxnLevel}
\title{Plots an interactive upset plot of overlapping input compounds at reaction class level 1}
\usage{
plotAnalyteOverlapPerRxnLevel(reactionsResults, includeCofactorMets = FALSE)
}
\arguments{
\item{reactionsResults}{output of getReactionsForAnalytes()}

\item{includeCofactorMets}{include metabolites labeled at cofactors within ChEBI (Default = FALSE)}
}
\value{
An interactive HTML upset plot that allows the user to visualize the overlap in the number of 
input compounds across level 1 of reaction classes.
}
\description{
Plots an interactive upset plot of overlapping input compounds at reaction class level 1
}
\examples{
\dontrun{
analytes.of.interest = c('chebi:58115', 'chebi:456215', 'chebi:58245', 'chebi:58450',
'chebi:17596', 'chebi:16335', 'chebi:16750', 'chebi:172878',
'chebi:62286', 'chebi:77897', 'uniprot:P30566','uniprot:P30520',
'uniprot:P00568', 'uniprot:P23109', 'uniprot:P22102', 'uniprot:P15531')

reactionsLists <- RaMP::getReactionsForAnalytes(analytes = analytes.of.interest, 
    includeTransportRxns = F, humanProtein = T, db = rampDB)

plotAnalyteOverlapPerRxnLevel(reactionsLists)
}

}
