% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotCataNetwork}
\alias{plotCataNetwork}
\title{Plots a network based on gene-metabolite relationships}
\usage{
plotCataNetwork(catalyzeDf = "")
}
\arguments{
\item{catalyzeDf}{a data.frame output by rampFastCata() that contains analytes that are in the same reaction}
}
\value{
An interactive HTML plot that allows the user to pan/zoom into regions of interest. User genes/
metabolites are highlighted in blue, whereas analytes found by the function are orange.
}
\description{
Plots a network based on gene-metabolite relationships
}
\examples{
\dontrun{
inputs.of.interest <- c("kegg:C00186" , "hmdb:HMDB0000148", "kegg:C00780", "hmdb:HMDB0000064",
      "ensembl:ENSG00000115850", "uniprot:Q99259")

new.transcripts <- rampFastCata(analytes = inputs.of.interest, db = rampDB)

plotCataNetwork(head(new.transcripts$HMDB_Analyte_Associations, n=100))
}
}
