% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab4.R
\name{plotCataNetwork}
\alias{plotCataNetwork}
\title{Plots a network based on gene-metabolite relationships}
\usage{
plotCataNetwork(catalyzedf = NULL)
}
\arguments{
\item{catalyzedf}{a data.frame output by rampFastCata() that contains analytes that are in the same reaction}
}
\value{
a list of nodes and edges
}
\description{
Plots a network based on gene-metabolite relationships
}
\examples{
\dontrun{
catalyzedf <- rampFastCata(analytes="creatine",conpass="mypassword",NamesOrIds="names")
plotCataNetwork(catalyzedf)
}
}
