% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab4.R
\name{plotCataNetwork}
\alias{plotCataNetwork}
\title{Plots a network based on gene-metabolite relationships}
\usage{
plotCataNetwork(catalyzedf = NULL)
}
\arguments{
\item{catalyzedf}{a data.frame output by rampFastCata() that contains analytes that are in the same reaction}
}
\value{
An interactive HTML plot that allows the user to pan/zoom into regions of interest. User genes/metabolites are highlighted in blue, whereas analytes found by the function are orange.
}
\description{
Plots a network based on gene-metabolite relationships
}
\examples{
\dontrun{
catalyzedf <- rampFastCata(analytes="creatine",conpass="mypassword",NamesOrIds="names")
plotCataNetwork(catalyzedf)
}
}
