% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab4.R
\name{rampFastCata}
\alias{rampFastCata}
\title{Return analytes that has catalyzation relation with given analyte}
\usage{
rampFastCata(analytes, conpass = NULL, dbname = "ramp", username = "root",
  host = "localhost")
}
\arguments{
\item{analytes}{a vector of analytes (genes or metabolites) that need to be searched}

\item{conpass}{password for database access (string)}

\item{dbname}{name of the mysql database (default is "ramp")}

\item{username}{username for database access (default is "root")}

\item{host}{host name for database access (default is "localhost")}
}
\value{
a data.frame that contains search results
}
\description{
Return analytes that has catalyzation relation with given analyte
}
