% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnGeneMetab_SameRxn.R
\name{rampFastCata}
\alias{rampFastCata}
\title{Retrieves analytes that involved in same reaction as input metabolite}
\usage{
rampFastCata(analytes = "none", namesOrIds = "ids", db = RaMP())
}
\arguments{
\item{analytes}{a vector of analytes that need to be searched}

\item{namesOrIds}{whether input is "names" or "ids" (default is "ids")}

\item{db}{a RaMP database object}
}
\value{
a list of two dataframes containing query results from HMDB and Rhea. If the input is a metabolite, the function will output
gene transcript common names and source IDs that are known to catalyze
reactions in the same pathway as that metabolite. Conversely, if the input
is a gene, the function will return the common name and source id of metabolites
known to be catalyzed directly or indirectly by that gene. Input ids and common names will be returned.
If no input ids or names are found in the database, the return value will be an empty data frame, 0 rows.
}
\description{
Retrieves analytes that involved in same reaction as input metabolite
}
\examples{
\dontrun{
inputs.of.interest <- c("kegg:C00186" , "hmdb:HMDB0000148", "kegg:C00780",
 "hmdb:HMDB0000064", "ensembl:ENSG00000115850", "uniprot:Q99259")

new.transcripts <- rampFastCata( analytes = inputs.of.interest, db = rampDB )
}
}
