% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnGeneMetab_SameRxn.R
\name{rampFastCata}
\alias{rampFastCata}
\title{Retrieves analytes that involved in same reaction as input metabolite}
\usage{
rampFastCata(analytes = NULL, NameOrIds = "ids")
}
\arguments{
\item{analytes}{a vector of analytes that need to be searched}

\item{NameOrIds}{whether input is "names" or "ids" (default is "ids")}
}
\value{
a dataframe containing query results. If the input is a metabolite, the function will output
gene transcript common names and source IDs that are known to catalyze
reactions in the same pathway as that metabolite. Conversely, if the input
is a gene, the function will return the common name and source id of metabolites
known to be catalyzed directly or indirectly by that gene.
}
\description{
Retrieves analytes that involved in same reaction as input metabolite
}
\examples{
\dontrun{
pkg.globals <- setConnectionToRaMP(dbname="ramp2",username="root",conpass="",host = "localhost")
rampFastCata(analytes="creatine",NameOrIds="names")
}
}
