% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab3.R
\name{rampFastPathFromMeta}
\alias{rampFastPathFromMeta}
\title{Fast search given a list of metabolites}
\usage{
rampFastPathFromMeta(synonym, find_synonym = FALSE, conpass = NULL,
  dbname = "ramp", username = "root", synonymOrIdS = "ids")
}
\arguments{
\item{synonym}{a vector of synonym that need to be searched}

\item{find_synonym}{bool if find all synonyms or just return same synonym}

\item{conpass}{password for database access (string)}

\item{dbname}{name of the mysql database (default is "ramp")}

\item{username}{username for database access (default is "root")}

\item{synonymOrIdS}{whether to return "synonyms" or "ids" (default is "ids")}
}
\value{
a list contains all metabolits as name and pathway inside.

Apply famil function...
}
\description{
Fast search given a list of metabolites
}
