% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab3.R
\name{rampFastPathFromMeta}
\alias{rampFastPathFromMeta}
\title{Fast search given a list of metabolites}
\usage{
rampFastPathFromMeta(synonym, find_synonym = FALSE, con)
}
\arguments{
\item{synonym}{a vector of synonym that need to be searched}

\item{find_synonym}{bool if find all synonyms or just return same synonym}

\item{con}{a connection object returned from the function connectToRaMP()}
}
\value{
a list contains all metabolits as name and pathway inside.

Apply famil function...
}
\description{
Fast search given a list of metabolites
}
