% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab3.R
\name{rampFastPathFromMeta}
\alias{rampFastPathFromMeta}
\title{tPathFromMetast search given a list of metabolites}
\usage{
rampFastPathFromMeta(analytes, find_synonym = FALSE, conpass = NULL,
  dbname = "ramp", username = "root", NameOrIds = "ids")
}
\arguments{
\item{analytes}{a vector of analytes (genes or metabolites) that need to be searched}

\item{find_synonym}{find all synonyms or just return same synonym (T/F)}

\item{conpass}{password for database access (string)}

\item{dbname}{name of the mysql database (default is "ramp")}

\item{username}{username for database access (default is "root")}

\item{NameOrIds}{whether input is "names" or "ids" (default is "ids")}
}
\value{
a list contains all metabolits as name and pathway inside.
}
\description{
tPathFromMetast search given a list of metabolites
}
