% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{rampFastPathFromSource}
\alias{rampFastPathFromSource}
\title{Fast search given a list of metabolites source Id}
\usage{
rampFastPathFromSource(sourceid, find_synonym = FALSE)
}
\arguments{
\item{sourceid}{a vector of synonym that need to be searched}
}
\value{
a list contains all metabolits as name and pathway inside.

Apply famil function...
}
\description{
Fast search given a list of metabolites source Id
}
