% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab4.R
\name{rampFileOfAnalytes_tab4}
\alias{rampFileOfAnalytes_tab4}
\title{Generate data.frame from given files}
\usage{
rampFileOfAnalytes_tab4(infile, conpass = NULL, dbname = "ramp",
  username = "root", host = "localhost")
}
\arguments{
\item{infile}{a file object given from files}

\item{conpass}{password for database access (string)}

\item{dbname}{name of the mysql database (default is "ramp")}

\item{username}{username for database access (default is "root")}

\item{host}{host name for database access (default is "localhost")}
}
\value{
a data.frame either from multiple csv file
}
\description{
identifing the file type, then it returns table output to 
shiny renderTable function as preview of searching data
}
