% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab4.R
\name{rampFileOfAnalytes_tab4}
\alias{rampFileOfAnalytes_tab4}
\title{Generate dataframe from given files for shiny app input list of metabolites}
\usage{
rampFileOfAnalytes_tab4(
  infile,
  conpass = NULL,
  dbname = "ramp",
  username = "root",
  host = "localhost"
)
}
\arguments{
\item{infile}{a file object given from files}

\item{conpass}{password for database access (string)}

\item{dbname}{name of the mysql database (default is "ramp")}

\item{username}{username for database access (default is "root")}

\item{host}{host name for database access (default is "localhost")}
}
\value{
a dataframe either from multiple csv file
}
\description{
Generate dataframe from given files for shiny app input list of metabolites
}
