% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{rampFindSourceFromId}
\alias{rampFindSourceFromId}
\title{Find all source from given list of RaMP Ids}
\usage{
rampFindSourceFromId(rampId = NULL, full = T)
}
\arguments{
\item{rampId}{could be a data frame return by rampFindSynonymFromSynonym
containing all information related to synonym. Or can be a list of 
rampId
@param full return whole searching result or not.}
}
\description{
Find all source from given list of RaMP Ids
}
