% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{rampFindSourceRampId}
\alias{rampFindSourceRampId}
\title{Find rampId from given source ID
The rampId can be plugged in other functions to continue query}
\usage{
rampFindSourceRampId(sourceId)
}
\arguments{
\item{sourceId}{a data frame or string separated by comma or string 
separated by new line}
}
\value{
data.frame that has sourceId and rampId and source as columns
}
\description{
Find rampId from given source ID
The rampId can be plugged in other functions to continue query
}
