% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{rampFindSynonymFromSynonym}
\alias{rampFindSynonymFromSynonym}
\title{Find all synonym from a given metabolite's name
This function is used to filter out some super common synonyms like glyceride
Now, this function only format the user input, so the user vector, dataframe,
and entire string separated by comma are working.}
\usage{
rampFindSynonymFromSynonym(
  synonym,
  full = FALSE,
  return_rampIds = FALSE,
  db = RaMP()
)
}
\arguments{
\item{synonym}{name to search for}

\item{full}{bool if return whole data.frame}

\item{return_rampIds}{bool to return ramp Ids with output
(there are some common synonyms that will mess up whole searching)}

\item{db}{a RaMP database object}
}
\value{
a data frame that contains synonym in the first column rampId in the second column
}
\description{
Find all synonym from a given metabolite's name
This function is used to filter out some super common synonyms like glyceride
Now, this function only format the user input, so the user vector, dataframe,
and entire string separated by comma are working.
}
