% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab6.R
\name{rampFindSynonymFromSynonym}
\alias{rampFindSynonymFromSynonym}
\title{Find all synonym from a given metabolite's name}
\usage{
rampFindSynonymFromSynonym(synonym, full = F, find_synonym = F)
}
\arguments{
\item{full}{bool if return whole data.frame}

\item{find_synonym}{bool if find all synonyms or just return same synonym
as input (there are some common synonyms that will mess up whole searching)}
}
\description{
Find all synonym from a given metabolite's name
}
