% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampQueryHelper.R
\name{rampFindSynonymFromSynonym}
\alias{rampFindSynonymFromSynonym}
\title{Find all synonym from a given metabolite's name
This function is used to filter out some super common synonyms like glyceride
Now, this function only format the user input, so the user vector, dataframe,
and entire string separated by comma are working.}
\usage{
rampFindSynonymFromSynonym(synonym, full = FALSE, find_synonym = FALSE,
  conpass = NULL, dbname = "ramp", username = "root")
}
\arguments{
\item{synonym}{name to search for}

\item{full}{bool if return whole data.frame}

\item{find_synonym}{bool if find all synonyms or just return same synonym}

\item{conpass}{password for database access (string)}

\item{dbname}{name of the mysql database (default is "ramp")}

\item{username}{username for database access (default is "root")
as input (there are some common synonyms that will mess up whole searching)}
}
\description{
Find all synonym from a given metabolite's name
This function is used to filter out some super common synonyms like glyceride
Now, this function only format the user input, so the user vector, dataframe,
and entire string separated by comma are working.
}
