% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab3.R
\name{rampFisherTestData}
\alias{rampFisherTestData}
\title{Calculate fisher test p-values from pathways returned when querying a list of genes
or metabolites (output of rampFastPathFromMeta)}
\usage{
rampFisherTestData(rampOut, analyte_type = "metabolites",
  total_analytes = 500, conpass = NULL, dbname = "ramp",
  username = "root")
}
\arguments{
\item{rampOut}{The data frame generated by rampFastPathFromMeta}

\item{analyte_type}{specify whether to do test on 'metabolites' or 'genes'}

\item{total_analytes}{number of total genes or metabolites analyzed in the experiment (e.g. background) (default is 500, with assumption that analyte_type is "metabolite")}

\item{conpass}{password for database access (string)}

\item{dbname}{name of the mysql database (default is "ramp")}

\item{username}{username for database access (default is "root")}
}
\value{
a data frame with three columns pathwayID, Number in pathway,
number out of pathway
}
\description{
Calculate fisher test p-values from pathways returned when querying a list of genes
or metabolites (output of rampFastPathFromMeta)
}
