% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab3.R
\name{rampGenerateBarPlot}
\alias{rampGenerateBarPlot}
\title{Reformat the result of query (get pathways from analyte(s)) for input into barplot
function}
\usage{
rampGenerateBarPlot(df)
}
\arguments{
\item{df}{A dataframe that has information for bar plot}
}
\value{
a list with the analyte names for each pathway that is represented in the list
}
\description{
each pathway contains all metabolites inside that pathway
}
