% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab3.R
\name{rampOneFisherTest}
\alias{rampOneFisherTest}
\title{Do fisher test for only one pathway from search result
clicked on highchart}
\usage{
rampOneFisherTest(pathwaydf, total_analytes = 500,
  analyte_type = "metabolites", conpass = NULL, dbname = "ramp",
  username = "root")
}
\arguments{
\item{pathwaydf}{a data frame resulting from rampFastPathFromMeta}

\item{total_analytes}{number of total genes or metabolites analyzed in the experiment (e.g. background) (default is 500, with assumption that analyte_type is "metabolite")}

\item{analyte_type}{"metabolites" or "genes" (default is "metabolites")}

\item{conpass}{password for database access (string)}

\item{dbname}{name of the mysql database (default is "ramp")}

\item{username}{username for database access (default is "root")}
}
\description{
Do fisher test for only one pathway from search result
clicked on highchart
}
