% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab3.R
\name{rampOneFisherTest}
\alias{rampOneFisherTest}
\title{Do fisher test for only one pathway from search result
clicked on highchart}
\usage{
rampOneFisherTest(pathway_list, pathway, num_of_meta, conpass = NULL,
  dbname = "ramp", username = "root")
}
\arguments{
\item{pathway_list}{a vector of all pathways searched from given metabolites}

\item{pathway}{single pathway name that is applied by Fisher test}

\item{num_of_meta}{integer that describe how many initial metabolites}

\item{conpass}{password for database access (string)}

\item{dbname}{name of the mysql database (default is "ramp")}

\item{username}{username for database access (default is "root")}
}
\description{
Do fisher test for only one pathway from search result
clicked on highchart
}
