% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralFunctions.R
\name{rampPathFromMeta}
\alias{rampPathFromMeta}
\title{Send query to database that find pathway}
\usage{
rampPathFromMeta(synonym, maxItems, con)
}
\arguments{
\item{synonym}{the string that describe given synonym name.}

\item{maxItems}{the integer that limit maximum returned by}

\item{con}{a connection object returned from the function connectToRaMP()
query}
}
\value{
a dataframe that has given synonym as column name
}
\description{
From given metabolite's synonym, it search through all databases to find
all pathways that have the metabolite.
}
