% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab1.R
\name{rampReadQuery}
\alias{rampReadQuery}
\title{Function that sends query with tryCatch
Return the customized error messages}
\usage{
rampReadQuery(query)
}
\arguments{
\item{query}{a string that follows syntax of MySQL and is used
to query database.}
}
\description{
Function that sends query with tryCatch
Return the customized error messages
}
