% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab3.R
\name{runCombinedFisherTest}
\alias{runCombinedFisherTest}
\title{Do fisher test for only one pathway from search result
clicked on highchart}
\usage{
runCombinedFisherTest(
  pathwaydf,
  total_metabolites = NULL,
  total_genes = 20000,
  min_analyte = 2,
  con = NULL
)
}
\arguments{
\item{pathwaydf}{a data frame resulting from getPathwayFromAnalyte}

\item{total_metabolites}{number of metabolites analyzed in the experiment (e.g. background) (default is 1000; set to 'NULL' to retrieve total number of metabolites that map to any pathway in RaMP). Assumption that analyte_type is "metabolite")}

\item{total_genes}{number of genes analyzed in the experiment (e.g. background) (default is 20000, with assumption that analyte_type is "genes")}

\item{min_analyte}{if the number of analytes (gene or metabolite) in a pathway is
< min_analyte, do not report}

\item{con}{MySQL connection (MySQLConnection)}
}
\value{
a list containing two entries: [[1]] fishresults, a dataframe containing pathways with Fisher's p values (raw and with FDR and Holm adjustment), number of user analytes in pathway, total number of analytes in pathway, and pathway source ID/database. [[2]] analyte_type, a string specifying the type of analyte input into the function ("genes", "metabolites", or "both")
}
\description{
Do fisher test for only one pathway from search result
clicked on highchart
}
\examples{
\dontrun{
pathwaydf<-getPathwayFromAnalyte(c("MDM2","TP53","glutamate","creatinine"),
                NameOrIds="names", conpass=conpass)
fisher.results <- runCombinedFisherTest(pathwaydf=pathwaydf,conpass=conpass)
}
}
