% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnPathwaysEnrich_InputAnalytes.R
\name{runCombinedFisherTest}
\alias{runCombinedFisherTest}
\title{Do fisher test for only one pathway from search result
clicked on highchart}
\usage{
runCombinedFisherTest(
  analytes,
  namesOrIds = "ids",
  totalGenes = 20000,
  minAnalyte = 2,
  alternative = "less",
  minPathwaySize = 5,
  maxPathwaySize = 150,
  includeRaMPids = FALSE,
  backgroundType = "database",
  background = "database",
  pathwayDefinitions = "RaMP",
  includeSmpdb = FALSE,
  db = RaMP()
)
}
\arguments{
\item{analytes}{a vector of analytes (genes or metabolites) that need to be searched}

\item{namesOrIds}{whether input is "names" or "ids" (default is "ids", must be the same for analytes and background)}

\item{totalGenes}{number of genes analyzed in the experiment (e.g. background) (default is 20000, with assumption that analyteType is "genes")}

\item{minAnalyte}{if the number of analytes (gene or metabolite) in a pathway is
< minAnalyte, do not report}

\item{alternative}{alternative hypothesis test passed on to fisher.test().  Options are two.sided, greater, or less (default is "less")}

\item{minPathwaySize}{the minimum number of pathway members (genes and metabolites) to include the pathway in the output (default = 5)}

\item{maxPathwaySize}{the maximum number of pathway memnbers (genes and metaboltes) to include the pathway in the output (default = 150)}

\item{includeRaMPids}{include internal RaMP identifiers (default is "FALSE")}

\item{backgroundType}{type of background that is input by the user.  Opions are "database" if user wants all
analytes from the RaMP database to be used as background; "file", if user wnats to input a file path with a list of background
analytes; "list", if user wants to input a vector of analyte IDs; "biospecimen", if user wants to specify a
biospecimen type (e.g. blood, adipose tissue, etc.) and have those biospecimen-specific analytes used.  For genes,
only the "database" option is used.}

\item{background}{background to be used for Fisher's tests.  If parameter 'backgroundType="database"', this parameter
is ignored (default="database"); if parameter 'backgroundType= "file"', then 'background' should be a file name (with
directory); if 'backgroundType="list"', then 'background' should be a vector of RaMP IDs; if 'backgroud_type="biospecimen"'
then users should specify one of the following: "Blood", "Adipose tissue", "Heart", "Urine", "Brain", "Liver", "Kidney",
"Saliva", and "Feces"}

\item{pathwayDefinitions}{If "RaMP" (default), use pathway definitions within RaMP-DB. Else, supply path to gmx file containing custom pathway definitions. GMX files are a tab-separated format that contain one analyte set per column, with the name of the set in the first row, and constituent analytes in subsequent rows. Please supply a .xls or .xlsx file. If supplying pathway definitions for genes and metabolites, ensure that metabolite definitions are on tab 1, and gene definitions are on tab2.}

\item{includeSmpdb}{Include pathways from smpdb/hmdb in analysis. Excluded by default since definitions are highly redundant}

\item{db}{a RaMP databse object}
}
\value{
a list containing two entries: [[1]] fishresults, a dataframe containing pathways with Fisher's p values
(raw and with FDR and Holm adjustment), number of user analytes in pathway, total number of analytes in pathway,
and pathway source ID/database. [[2]] analyteType, a string specifying the type of analyte input into the function ("genes", "metabolites", or "both")
}
\description{
Do fisher test for only one pathway from search result
clicked on highchart
}
\examples{
\dontrun{
fisher.results <- runCombinedFisherTest(analytes = c(
"hmdb:HMDB0000033",
"hmdb:HMDB0000052",
"hmdb:HMDB0000094",
"hmdb:HMDB0000161",
"hmdb:HMDB0000168",
"hmdb:HMDB0000191",
"hmdb:HMDB0000201",
"chemspider:10026",
"hmdb:HMDB0006059",
"Chemspider:6405",
"CAS:5657-19-2",
"hmdb:HMDB0002511",
"chemspider:20171375", "CAS:133-32-4",
"CAS:5746-90-7",
"CAS:477251-67-5",
"hmdb:HMDB0000695",
"chebi:15934",
"CAS:838-07-3",
"hmdb:HMDBP00789",
"hmdb:HMDBP00283",
"hmdb:HMDBP00284",
"hmdb:HMDBP00850"
), db = rampDB )

fisher.results <- runCombinedFisherTest(analytes = analyte.list, namesOrIds = "ids")
}
}
