% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnPathwaysEnrich_InputAnalytes.R
\name{runFisherTest}
\alias{runFisherTest}
\title{Do fisher test for only one pathway from search result
clicked on highchart}
\usage{
runFisherTest(
  pathwaydf,
  total_metabolites = NULL,
  total_genes = 20000,
  analyte_type = "metabolites",
  MCall = T,
  alternative = "less"
)
}
\arguments{
\item{pathwaydf}{a data frame resulting from getPathwayFromAnalyte}

\item{total_metabolites}{number of metabolites analyzed in the experiment (e.g. background) (default is 1000; set to 'NULL' to retrieve total number of metabolites that map to any pathway in RaMP). Assumption that analyte_type is "metabolite")}

\item{total_genes}{number of genes analyzed in the experiment (e.g. background) (default is 20000, with assumption that analyte_type is "genes")}

\item{analyte_type}{"metabolites" or "genes" (default is "metabolites")}

\item{MCall}{T/F if true, all pathways are used for multiple comparison corrections; if false, only pathways covering user analytes will be used (default is "T")}

\item{alternative}{alternative hypothesis test passed on to fisher.test().  Options are two.sided, greater, or less (default is "less")}
}
\value{
a dataframe with columns containing pathway ID, fisher's p value, user analytes in pathway, and total analytes in pathway
}
\description{
Do fisher test for only one pathway from search result
clicked on highchart
}
