% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnPathwaysEnrich_InputAnalytes.R
\name{runFisherTest}
\alias{runFisherTest}
\title{Do fisher test for only one pathway from search result
clicked on highchart}
\usage{
runFisherTest(
  pathwaydf,
  backgrounddf = NULL,
  total_metabolites = NULL,
  total_genes = 20000,
  analyte_type = "metabolites",
  conpass = NULL,
  dbname = "ramp",
  username = "root",
  host = "localhost",
  socket = NULL,
  MCall = T,
  alternative = "less"
)
}
\arguments{
\item{pathwaydf}{a data frame resulting from getPathwayFromAnalyte}

\item{backgrounddf}{optional dataframe resulting from getPathwayFromAnalyte run on all analytes detected in study. This will be used as the background for the Fisher's contingency table. If left NULL, all analytes in RaMP are used as background}

\item{total_metabolites}{number of metabolites analyzed in the experiment (e.g. background) (default is 1000; set to 'NULL' to retrieve total number of metabolites that map to any pathway in RaMP). Assumption that analyte_type is "metabolite")}

\item{total_genes}{number of genes analyzed in the experiment (e.g. background) (default is 20000, with assumption that analyte_type is "genes")}

\item{analyte_type}{"metabolites" or "genes" (default is "metabolites")}

\item{conpass}{password for database access (string)}

\item{dbname}{name of the mysql database (default is "ramp")}

\item{username}{username for database access (default is "root")}

\item{host}{host name for database access (default is "localhost")}

\item{socket}{(optional) location of mysql.sock file}

\item{MCall}{T/F if true, all pathways are used for multiple comparison corrections; if false, only pathways covering user analytes will be used (default is "T")}

\item{alternative}{alternative hypothesis test passed on to fisher.test().  Options are two.sided, greater, or less (default is "less")}
}
\value{
a dataframe with columns containing pathway ID, fisher's p value, user analytes in pathway, and total analytes in pathway
}
\description{
Do fisher test for only one pathway from search result
clicked on highchart
}
