% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampFastQueryTab3.R
\name{runFisherTest}
\alias{runFisherTest}
\title{Do fisher test for only one pathway from search result
clicked on highchart}
\usage{
runFisherTest(pathwaydf, total_metabolites = NULL, total_genes = 20000,
  analyte_type = "metabolites", conpass = NULL, dbname = "ramp",
  username = "root")
}
\arguments{
\item{pathwaydf}{a data frame resulting from rampFastPathFromMeta}

\item{total_metabolites}{number of metabolites analyzed in the experiment (e.g. background) (default is total number of metabolites that map to a pathway in RaMP, with assumption that analyte_type is "metabolite")}

\item{total_genes}{number of genes analyzed in the experiment (e.g. background) (default is 20000, with assumption that analyte_type is "genes")}

\item{analyte_type}{"metabolites" or "genes" (default is "metabolites")}

\item{conpass}{password for database access (string)}

\item{dbname}{name of the mysql database (default is "ramp")}

\item{username}{username for database access (default is "root")}
}
\value{
a dataframe with pathway enrichment (based on Fisher's test) results
}
\description{
Do fisher test for only one pathway from search result
clicked on highchart
}
