% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbWorker.R
\name{runQuery}
\alias{runQuery}
\title{Utility wrapper method for queries handling mysql and sqlite}
\usage{
runQuery(sql)
}
\arguments{
\item{sql}{sql query to run}
}
\value{
returns the query result, Note: The returned data type is dependent on query
}
\description{
Utility wrapper method for queries handling mysql and sqlite
}
