% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRaMP.R
\name{runRaMPapp}
\alias{runRaMPapp}
\title{Run Shiny App}
\usage{
runRaMPapp(conpass = NULL)
}
\arguments{
\item{conpass}{password for database access (string)}
}
\description{
This function launches the RShiny app.  It requires a connection to the RaMP database as input, which requires running the function connectoToRaMP() and providing the MySQL password.
}
\examples{
\dontrun{
con <- connectToRaMP(dbname="ramp",username="root",password="mypassword")
runRaMPapp(con=con)
}
}
