% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRaMP.R
\name{runRaMPapp}
\alias{runRaMPapp}
\title{Run Shiny App}
\usage{
runRaMPapp(dbname = "ramp", username = "root", password = NULL,
  host = NULL, update = FALSE)
}
\arguments{
\item{dbname}{the name of the dadtabase (by default a string that is database name including all tables}

\item{username}{a string that is username for database}

\item{password}{a string this is password for database}

\item{host}{a string that stand for host}

\item{update}{a logic value determines if RaMP update local data}
}
\description{
This function connects to the RaMP database (provided within the package), and launches the RaMP web application.  The password for the mysql connection must be provided.
}
