% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectToRaMP.R
\name{setConnectionToRaMP}
\alias{setConnectionToRaMP}
\title{Set Connection Parameters or RaMP}
\usage{
setConnectionToRaMP(
  dbname = "ramp",
  username = "root",
  conpass = "",
  host = "localhost",
  socket = "",
  is_sqlite = F,
  sqlite_file_path = ""
)
}
\arguments{
\item{dbname}{the name of the database (by default a string that is database name including all tables}

\item{username}{a string that is username for database (default: root)}

\item{conpass}{password for database (string)}

\item{host}{a string that stand for host}

\item{socket}{optional, location of mySQL.sock file (useful when running RaMP on remote clusters)}
}
\description{
Set Connection Parameters or RaMP
}
\examples{
\dontrun{
pkg.globals <- setConnectionToRaMP(dbname="ramp2",username="root",conpass="",host = "localhost")
}
}
