% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processNewRamp.R
\name{updateOverlapMatrices}
\alias{updateOverlapMatrices}
\title{Update and save the overlap matrices based on current version of RaMP}
\usage{
updateOverlapMatrices(
  method,
  all,
  conpass,
  host = "localhost",
  dbname = "ramp",
  username = "root"
)
}
\arguments{
\item{method}{a string that specifies algorithm to compute overlap matrix
should be 'balanced' or 'weighted'}

\item{all}{a string that specifies which matrices to compute, should be in
'all','analyte'}

\item{conpass}{password for database access (string)}

\item{host}{host name for database access (default is "localhost")}

\item{dbname}{name of the mysql database (default is "ramp")}

\item{username}{username for database access (default is "root")}
}
\description{
Update and save the overlap matrices based on current version of RaMP
}
