% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writingFunctions.R
\name{write_FishersResults}
\alias{write_FishersResults}
\title{Function that writes Fishers Test results to a CSV file}
\usage{
write_FishersResults(fishResults = NULL, outputfile = NULL,
  clusters = NULL)
}
\arguments{
\item{fishResults}{a data frame returned by function runCombinedFisherTest()}

\item{outputfile}{name of output file}

\item{clusters}{optinal clusteres as returned from the function findCluster()}
}
\description{
Function that writes Fishers Test results to a CSV file
}
