% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writingFunctions.R
\name{write_FishersResults}
\alias{write_FishersResults}
\title{Function that writes Fishers Test results, after clustering to a CSV file}
\usage{
write_FishersResults(fishResults = NULL, outputfile = NULL)
}
\arguments{
\item{fishResults}{a data frame returned by function runCombinedFisherTest()}

\item{outputfile}{name of output file}
}
\description{
Function that writes Fishers Test results, after clustering to a CSV file
}
