% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writingFunctions.R
\name{writePathwaysToCSV}
\alias{writePathwaysToCSV}
\title{Function that writes output from getPathwayFromAnalyte() to a CSV file}
\usage{
writePathwaysToCSV(myPathways = "none", outputFile = "none")
}
\arguments{
\item{myPathways}{data frame returned by function getPathwayFromAnalyte()}

\item{outputFile}{name of output file}
}
\description{
Function that writes output from getPathwayFromAnalyte() to a CSV file
}
